% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{calculate_d18O_evap}
\alias{calculate_d18O_evap}
\title{Evaporation d18O}
\usage{
calculate_d18O_evap(atmp, ltmp, RH, d18O_pcpn, d18O_lake)
}
\arguments{
\item{atmp}{air temperature (K)}

\item{ltmp}{lake surface temperature (K)}

\item{RH}{relative humidity (\%)}

\item{d18O_pcpn}{d18O isotopic composition of precipitation}

\item{d18O_lake}{d18O isotopic composition of the lake}
}
\value{
d18O_evap - the isotope composition of evaporation
}
\description{
Calculates the isotope composition of evaporation based on equation 5 in
Krabbenhoft et al. (1990). Note that alpha* is equivalent to 1/alpha.
}
\references{
Krabbenhoft, D. P., C. J. Bowser, M. P. Anderson, and J. W.
  Valley. (1990). Estimating Groundwater Exchange with Lakes: 1. The Stable
  Isotope Mass Balance Method. Water Resources Research, 26(10):2445-2453.
  https://doi.org/10.1029/WR026i010p02445
}
