% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dV.R
\name{calculate_dV}
\alias{calculate_dV}
\title{Change in Lake Volume}
\usage{
calculate_dV(stage_vol, daily_stage = NULL, as_depth = TRUE,
  tmp_depth_fix = FALSE)
}
\arguments{
\item{stage_vol, }{a data frame with the following columns:
\itemize{
\item stage_m - stage of lake (m) above mean sea level
\item surf_area_m2 - surface area of lake (m^2) at given elevation
\item volume_m3 - volume of lake (m^3) at given elevation
}}

\item{daily_stage}{a data frame with the following columns:
\itemize{
\item date - date of lake stage measurement
\item stage_m - daily measurement of lake level (m)
}}

\item{as_depth}{defaults to TRUE to normalize volume (m^3) by average surface
area of lake during the time segment (m^2) and return change
in volume as a depth (mm). If FALSE, returns change in lake
volume as a volume (m^3).}

\item{tmp_depth_fix}{defaults to FALSE. If TRUE, calculates change in volume
as a depth based solely on change in stage (mm).}
}
\value{
dV, the change in lake volume during this time period as a volume
        (m^3) or depth (mm)
}
\description{
Calculates the change in lake volume over the course of the inputed time
series of daily lake stage measurements (m).
}
