% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GW_functions.R
\name{calculate_GW_inflow}
\alias{calculate_GW_inflow}
\title{Groundwater Inflow}
\usage{
calculate_GW_inflow(P, E, d18O_pcpn, d18O_lake, d18O_GWin, d18O_evap)
}
\arguments{
\item{P}{precipitation (mm)}

\item{E}{evapotranspiration (mm)}

\item{d18O_pcpn}{d18O isotopic composition of precipitation}

\item{d18O_lake}{d18O isotopic composition of the lake}

\item{d18O_GWin}{d18O isotopic composition of groundwater inflow}

\item{d18O_evap}{d18O isotopic composition of evaporation}
}
\value{
GWin - the groundwtaer inflow to the lake (mm)
}
\description{
Calculates the groundwater inflow to a lake based on equation 4 in
Krabbenhoft et al. (1990).
}
\references{
Krabbenhoft, D. P., C. J. Bowser, M. P. Anderson, and J. W.
  Valley. (1990). Estimating Groundwater Exchange with Lakes: 1. The Stable
  Isotope Mass Balance Method. Water Resources Research, 26(10):2445-2453.
  https://doi.org/10.1029/WR026i010p02445
}
