% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_GW_outflow.R
\name{calculate_GW_outflow}
\alias{calculate_GW_outflow}
\title{Groundwater Outflow}
\usage{
calculate_GW_outflow(P, E, GWin, dVdt = 0)
}
\arguments{
\item{P}{precipitation (mm)}

\item{E}{evapotranspiration (mm)}

\item{GWin}{groundwater inflow (mm)}

\item{dVdt}{change in lake volume storage (mm)}
}
\value{
GWout - the groundwtaer outflow from the lake (mm)
}
\description{
Calculates the groundwater outflow as the remaining term in a water balance.
}
