% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_h2o_bal.R
\name{calculate_h2o_bal}
\alias{calculate_h2o_bal}
\title{Calculate Water Balance}
\usage{
calculate_h2o_bal(inputs)
}
\arguments{
\item{inputs}{a data frame with the date, air temperature (in degrees
C and Kelvin), relative humidity (percent),
precipitation (mm), lake evaporation (m), lake surface
temperature (Kelvin), change in lake volume (mm), d18O
isotopic composition of the lake, d18O isotopic
composition of precipitation, d18O isotopic composition
of GWin, d18O isotopic composition of GWout, d18O
isotopic composition of evaporation, and the well sites
used in each d18O_GWin and d18O_GWout mean. As output
by \code{\link{summarise_inputs}}.}
}
\value{
h2o_bal, the caculated water balance fluxes including precipitation,
        lake evaporation, change in lake volume, groundwater inflow, and
        groundwater outflow.
}
\description{
Give all required inputs, calculates the remaining terms of the water
balance: groundwater inflow and groundwater outflow.
}
\seealso{
\code{\link{summarise_inputs}}
}
