% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lake_evap.R
\name{calculate_lake_evap}
\alias{calculate_lake_evap}
\title{Calculate Lake Evaporation}
\usage{
calculate_lake_evap(weather, lst, stage_vol, site_dictionary, Lz = 90,
  lake_albedo = 0.08, wind_elev = 3, z0 = 0.02)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including date, atmp (air
temperature, deg C), RH (relative humidity, percent), P
(precipitation, mm), Rs (incoming solar radiation, MJ/m^2),
and wind (wind speed, m/s) as formatted in the
\code{\link{weather}} dataset.}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the \code{\link{lst}} dataset, subset
for a single lake.}

\item{stage_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the \code{\link{stage_vol}} dataset, subset
for a single lake.}

\item{site_dictionary}{a data frame with the lake, obs_type, site_id,
SWIMS_station, USGS_id, WBIC, and static_iso_type of
measurement sites, as in the
\code{\link{site_dictionary}} dataset, subset for a
single lake.}

\item{Lz}{the longitude of the local timezone (degrees west of Greenwich,
ranges from 0 to 360 degrees). Defaults to 90 for Central Time
Zone, USA.}

\item{lake_albedo}{albedo of the lake, defaults to 0.08 for open water.}

\item{wind_elev}{height of wind measurement, defaults to 3m for Hancock
station.}

\item{z0}{aeorodynamic roughness of landcover at site of weather measurements
(m), defaults to 0.02 for a grass.}
}
\value{
weather, a list with daily weather information including:
\describe{
\item{date}{day of each weather observation}
\item{atmp_min}{minimum air temperature for the day (deg C)}
\item{atmp_max}{maximum air temperature for the day (deg C)}
\item{RH_min}{minimum relative humidity for the day (percent)}
\item{RH_max}{maximum relative humidity for the day (percent)}
\item{P}{total precipitation for the day (mm)}
\item{ET}{total lake evaporation for the day (mm)}
}
}
\description{
Calculates Lake Evaporation
}
