% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_iso_site_functions.R
\name{classify_iso_site_static}
\alias{classify_iso_site_static}
\title{Classify Stable Isotope Measurements (Static)}
\usage{
classify_iso_site_static(isotopes, site_dictionary)
}
\arguments{
\item{isotopes}{a data frame with isotope measurements as formatted in the
\code{\link{isotopes}} dataset, subset to records for the lake
of interest.}

\item{site_dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
\code{\link{site_dictionary}} dataset, subset to
records for the lake of interest.}
}
\value{
isotopes - the same data frame provided to the function, but with an
                   additional column for "site_type" with values of
                   "precipitation", "lake", "upgradient", "downgradient", or
                   "".
}
\description{
Classifies the site type of isotope measurements as precipitation, lake,
upgradient, or downgradient. Upgradient vs. downgradient wells are determined
by static definitions in the site dictionary.
}
