% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_lake_isotopes.R
\name{classify_lake_isotopes}
\alias{classify_lake_isotopes}
\title{Classify Stable Isotope Measurements}
\usage{
classify_lake_isotopes(isotopes, water_levels, lake,
  site_file = "csls_site_dictionary.csv", filedir = "system.file",
  Xday = 7)
}
\arguments{
\item{isotopes}{a data frame with the following columns:}

\item{water_levels}{a data frame with the following columns:}

\item{lake}{name of lake to analyze (e.g., Pleasant, Long, or Plainfield).
This name corresponds to lake names in the site dictionary "lake"
column.}

\item{site_file}{filname of site dictionary with the following columns:}

\item{filedir}{name of directory with both csv files, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}

\item{Xday}{number of days to consider when calculating average water level
before isotope measurements, defaults to 7 days.}
}
\value{
isotopes - the same data frame provided to the function, but with an
                   additional column for "site_type" with values of
                   "precipitation", "lake", "upgradient", or "downgradient".
}
\description{
Classifies the site type of isotope measurements as precipitaiton, lake,
upgradient, or downgradient. Upgradient vs. downgradient wells are determined
by comparing the 7-day mean lake level and well level leading up to isotope
measurements.
}
