% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap}
\alias{d18O_evap}
\title{Evaporation d18O}
\usage{
d18O_evap(atmp, ltmp, RH, d18O_pcpn, d18O_lake)
}
\arguments{
\item{atmp}{air temperature (K)}

\item{ltmp}{lake surface temperature (K)}

\item{RH}{relative humidity (\%)}

\item{d18O_pcpn}{d18O isotopic composition of precipitation}

\item{d18O_lake}{d18O isotopic composition of the lake}
}
\value{
d18O_evap - the isotope composition of evaporation
}
\description{
Calculates the isotope composition of evaporation.
}
