% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap_normalized_humidity}
\alias{d18O_evap_normalized_humidity}
\title{Normalized Relative Humidity}
\usage{
d18O_evap_normalized_humidity(RH, es_a, es_l)
}
\arguments{
\item{RH}{relative humidity (\%)}

\item{es_a}{saturation vapor pressure for the air}

\item{es_l}{saturation vapor pressure for the lake}
}
\value{
h - the relative humidity normalized to the temperature of the
        surface water
}
\description{
Calculates the relative humidity normalized to the temperature of the surface
water
}
