% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap_sat_vapor_press}
\alias{d18O_evap_sat_vapor_press}
\title{Saturation Vapor Pressure}
\usage{
d18O_evap_sat_vapor_press(tmp)
}
\arguments{
\item{tmp}{temperature of air or water (K)}
}
\value{
es - saturation vapor pressure
}
\description{
Calculates the saturation vapor pressure based on temperature (of air or
water)
}
