% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap_total_frac}
\alias{d18O_evap_total_frac}
\title{Total Fractionation Factor}
\usage{
d18O_evap_total_frac(alpha, delta_epsilon)
}
\arguments{
\item{alpha}{equilibrium isotope fractionation factor at the temperature of
the air-water interface}

\item{delta_epsilon}{kinetic fractionation factor}
}
\value{
epsilon - the total fractionation factor
}
\description{
Calculates the total fractionation factor
}
