% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_monthly_weather.R
\name{daily_to_monthly_weather}
\alias{daily_to_monthly_weather}
\title{Daily to Monthly Weather}
\usage{
daily_to_monthly_weather(filename, filedir = "data", skip_lines = 7)
}
\arguments{
\item{filename}{name of daily weather file which includes:
\itemize{
\item date
\item atmp_max - maximum air temperature (deg C)
\item atmp_min - minimum air temperature (deg C)
\item relh_max - maximum relative humidity (%)
\item relh_min - minimum relative humidity (%)
\item pcpn - precipitation (mm)
\item rpet - reference potential evapotranspiration (mm)
}}

\item{filedir}{directory in which daily weather csv file resides}

\item{skip_lines}{number of rows in csv to skip before reading, default: 7}
}
\description{
This function takes daily max air temperature, min air temperature, max
relative humidity, min relative humidity, precipitaiton, and reference 
evapotranspiration as formatted by the Hancock, WI weather station and 
summarizes it at a monthly time step.
}
