% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_functions.R
\name{fill_timeseries_gaps}
\alias{fill_timeseries_gaps}
\title{Fill Gaps in Summary Data}
\usage{
fill_timeseries_gaps(df, timeseries)
}
\arguments{
\item{df}{data frame with a "date" column}

\item{timeseries}{vector with complete list of dates that should be in the
"date" column of df}
}
\value{
df, the same data frame with blank rows added to fill in timeseries,
       then sorted by date.
}
\description{
This function compares dates in a given (complete) timeseries and dataframe,
then adds blank rows for the dates that are missing in the dataframe.
}
