% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_functions.R
\name{find_nmonths}
\alias{find_nmonths}
\title{Identify Number of Months in Timeseries}
\usage{
find_nmonths(start_date, end_date)
}
\arguments{
\item{start_date}{date of first month in time series (assumes day = first day
of that month)}

\item{end_date}{date of last month in time series (assumes day = last day of
that month)}
}
\value{
nmonths - number of months in this time series
}
\description{
Calculates the number of months in a timeseries given a start and end date.
}
