% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_functions.R
\name{find_overlap_timeseries}
\alias{find_overlap_timeseries}
\title{Identify Months With Overlapping Data}
\usage{
find_overlap_timeseries(weather, lst, isotopes, lake_levels, gw_levels)
}
\arguments{
\item{weather}{a data frame with daily weather in the following columns:
\itemize{
\item date - date and time of weather observation
\item atmp - air temperature (deg C)
\item pcpn - precipitation (mm)
\item relh - relative humidity (percent)
\item rpet - reference potential evapotranspiration (mm)
}}

\item{lst}{a data frame with every-other-week lake surface temperature info:
\itemize{
\item WBIC - Water Body Identification Code for lake
\item date - date and time of observation
\item ltmp - lake surface temperature (degrees C)
\item units - units of lake surface temperature, should all be "DEGREES C"
}}

\item{isotopes}{a data frame with isotope measurements structured as follows:
\itemize{
\item date - date of measurement
\item lake - lake associated with measurement
\item site_id - unique ID for site of measurement, e.g. "PRECIP", "LONG",
               "LL-01"
\item d18O - isotopic composition for 18O (per mil)
\item d2H - isotopic composition for duterium (per mil)
}}

\item{lake_levels}{a data frame with daily lake levels structured as follows:
\itemize{
\item date - date of water level measurement
\item site_no - UGSG site number
\item obs_type - type of observation (LK = lake level)
\item level_m - water level in meters above mean sea level
}}

\item{gw_levels}{a data frame with daily groundwater levels structured as
                 follows:
\itemize{
\item date - date of water level measurement
\item site_no - UGSG site number
\item obs_type - type of observation (GW = groundwater level)
\item level_m - water level in meters above mean sea level
}}
}
\value{
a list with the following variables:
\describe{
\item{start_date}{start date (1st of the month) of timeseries with
                  overlapping input data}
\item{end_date}{end date (last day of the month) of timeseries with
                overlapping input data}
\item{nmonths}{number of months in timeseries with overlapping input data}
}
}
\description{
Identifies the start date, end date, and number of months in the monthly
timeseries with complete input data coverage.
}
