% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_functions.R
\name{find_start_date}
\alias{find_start_date}
\title{Identify Start Date of Timeseries}
\usage{
find_start_date(date_vector, all_days = FALSE)
}
\arguments{
\item{date_vector}{vector with dates in timeseries}

\item{all_days}{logical defaults to FALSE to indicate that timeseries does
not need to include values for every day within the month,
set to TRUE to force this behavior.}
}
\value{
\describe{
\item{start_date}{first useable date in the timeseries (rounded to the first
of the month)}
}
}
\description{
Identifies the starting date of a monthly timeseries (forced to the 1st of
the month)
}
