% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_lake_info.R
\name{format_lake_info}
\alias{format_lake_info}
\title{Format Lake Size & Depth Info for Lake Evaporation}
\usage{
format_lake_info(stage_vol)
}
\arguments{
\item{stage_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the \code{\link{stage_vol}} dataset, subset
for a single lake.}
}
\value{
lake, a list with lake data that includes:
\describe{
\item{A}{surface area of the lake (km^2).}
\item{depth_m}{depth of the lake (m)}
}
}
\description{
Extracts surface area and depth of lake and re-formats for input to lake
evaporation functions.
}
\seealso{
\code{\link{retrieve_csls_stage_volume}},
         \code{\link{stage_vol}}
}
