% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_lake_loc.R
\name{format_lake_loc}
\alias{format_lake_loc}
\title{Format Lake Location Info for Lake Evaporation}
\usage{
format_lake_loc(site_dictionary, Lz = 90)
}
\arguments{
\item{site_dictionary}{a data frame with the elev_m, lat_deg, and long_deg of
measurement sites, as in the
site_dictionary dataset, subset for a
single lake.}

\item{Lz}{the longitude of the local timezone (degrees west of Greenwich,
ranges from 0 to 360 degrees). Defaults to 90 for Central Time
Zone, USA.}
}
\value{
loc, a list with the following location-specific parameters:
\describe{
\item{z}{elevation above mean sea level (m)}
\item{phi}{latitude (radians). Positive for northern hemisphere, negative for
           southern hemisphere.}
\item{Lm}{longitude of location (degrees west of Greenwich)}
\item{Lz}{longitude of location's timezone (degrees west of Greenwich)}
}
}
\description{
Extracts location information for the lake from the site dictionary and
reformats for input into lake evaporation functions.
}
