% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_lake_weather.R
\name{format_lake_weather}
\alias{format_lake_weather}
\title{Format Lake Weather Info for Lake Evaporation}
\usage{
format_lake_weather(weather, day_zero, wind_elev = 3, z0 = 0.02)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including including date
(POSIXct), atmp (air temperature, deg C), P (precipitation, mm), RH
(relative humidity, percent), Rs (solar radiation, MJ), and wind (wind
speed, m/s)}

\item{day_zero}{day before the first day of weather info to retain}

\item{wind_elev}{height of wind measurement, defaults to 3m for Hancock
station.}

\item{z0}{aeorodynamic roughness of landcover at site of weather measurements
(m), defaults to 0.02 for a grass.}
}
\value{
weather, a list with daily weather data that includes:
\describe{
\item{datetimes}{datetimes of weather records [POSIXct]}
\item{P}{vector of daily precipitation (mm) corresponding with datetimes
         vector}
\item{atmp}{list with two vectors, "min" and "max", with mean daily min and
            max air temperature (degrees C) corresponding with datetimes
            vector}
\item{RH}{list with two vectors, "min" and "max", with mean daily min and
           max relative humidity (percent) corresponding with datetimes
           vector.}
\item{Rs}{vector of incoming solar or shortwave radiation (MJ/m^2/timestep),
          corresponding with datetimes vector.}
\item{wind}{vector with mean windspead (m/s), corresponding with datetimes
            vector}
\item{wind_elev}{height at which wind is measured (m)}
\item{dt}{string indicating the timestep of input weather series. Expects
          "hourly", "daily", or "monthly".}
\item{z0}{aerodynamic roughness of measurement site (m)}
 }
}
\description{
Takes hourly weather data, summarizes as daily weather data, and re-formats
for input to lake evaporation functions
}
