% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_timeseries_functions.R
\name{get_end_date}
\alias{get_end_date}
\title{Identify End Date}
\usage{
get_end_date(date_vector, all_days = FALSE)
}
\arguments{
\item{date_vector}{vector with dates in timeseries}

\item{all_days}{[logical] defaults to FALSE to indicate that timeseries does
not need to include values for every day within the month,
set to TRUE to force this behavior.}
}
\value{
end_date last useable date in the timeseries (rounded to the last day
        of the month)
}
\description{
Identifies the ending date of a monthly timeseries (forced to the last day of
the month)
}
