% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_values_functions.R
\name{get_monthly_d18O_evap}
\alias{get_monthly_d18O_evap}
\title{Monthly Evaporation d18O}
\usage{
get_monthly_d18O_evap(monthly_weather, monthly_lst, monthly_isotopes)
}
\arguments{
\item{monthly_weather}{a data frame as output by
\code{\link{get_monthly_weather}}}

\item{monthly_lst}{a data frame as output by \code{\link{get_monthly_lst}}}

\item{monthly_isotopes}{a data frame as output by
\code{\link{get_monthly_isotopes}}}
}
\value{
monthly_isotopes with d18O_evap column added
}
\description{
Calculates monthly Evaporation d18O based on monthly weather, lake surface
temperature, and isotope measurements.
}
