% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_dV.R
\name{get_monthly_dV}
\alias{get_monthly_dV}
\title{Monthly Change in Lake Volume}
\usage{
get_monthly_dV(lake_levels, month_info)
}
\arguments{
\item{lake_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to lake level records for the lake of interest.}

\item{month_info}{a list with with the start date, end date, and number of
months in the common timeseries of input data, as output by
\code{\link{get_overlap_months}}}
}
\value{
monthly_dV, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{dV}{change in lake volume during the month as a depth (mm)}
}
}
\description{
Calculates the monthly change in lake volume based on daily lake levels and a
stage-volume relationship from bathymetry analysis. .
}
