% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_h2o_bal.R
\name{get_monthly_h2o_bal}
\alias{get_monthly_h2o_bal}
\title{Monthly Water Balance}
\usage{
get_monthly_h2o_bal(monthly_weather, monthly_isotopes)
}
\arguments{
\item{monthly_weather}{data frame with the following columns:
\itemize{
\item date - first of the month for each monthly observation (Date)
\item atmp_K - mean monthly air temperature (K)
\item relh_pct - mean monthly relative humidity (percent)
\item pcpn_mm - monthly precipitaiton (mm)
\item rpet_mm - monthly reference potential evapotranspiration (mm)
}}

\item{monthly_isotopes}{a data frame with the following columns:
\itemize{
\item date - first of the month for each monthly observation (Date)
\item d18O_pcpn - mean precipitation stable isotope measurement for the month
\item d18O_lake - mean lake stable isotope measurement for the month
\item d18O_GWin - mean groundwater inflow stable isotope measurement for the
                 month
\item d18O_evap - evaporation stable isotope composition for the month
}}
}
\value{
lake_h2o_bal - a data frame with the following columns;
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{P}{monthly precipitation (mm)}
\item{E}{monthly evapotranspiration (mm)}
\item{GWin}{monthly groundwater inflow to the lake (mm)}
\item{GWout}{monthly groundwater outflow to the lake (mm)}
}
}
\description{
Calculates the water balance of a lake based measured fluxes and isotopic
signatures of precipitation, the lake, and inflowing groundwater wells. Based
on equations 2 and 4 in Krabbenhoft et al. (1990).
}
\references{
Krabbenhoft, D. P., C. J. Bowser, M. P. Anderson, and J. W.
  Valley. (1990). Estimating Groundwater Exchange with Lakes: 1. The Stable
  Isotope Mass Balance Method. Water Resources Research, 26(10):2445-2453.
  https://doi.org/10.1029/WR026i010p02445
}
