% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_weather.R
\name{get_monthly_weather}
\alias{get_monthly_weather}
\title{Monthly Weather}
\usage{
get_monthly_weather(weather)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including date, TMP (air
temperature, deg C), RH (relative humidity, percent), P
(precipitation, mm), and ET (evapotranspiration, mm) as
formatted in the \code{\link{weather}} dataset.}
}
\value{
monthly_weather, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{atmp_K}{mean monthly air temperature (K)}
\item{RH_pct}{mean monthly relative humidity (percent)}
\item{P_mm}{monthly precipitaiton (mm)}
\item{ET_mm}{monthly reference potential evapotranspiration (mm)}
}
}
\description{
Summarizes sub-monthly weather at a monthly timestep
}
\seealso{
\code{\link{retrieve_csls_weather}}, \code{\link{weather}}
}
