% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_timeseries_functions.R
\name{get_start_date}
\alias{get_start_date}
\title{Identify Start Date}
\usage{
get_start_date(date_vector, all_days = FALSE)
}
\arguments{
\item{date_vector}{vector with dates in timeseries}

\item{all_days}{[logical] defaults to FALSE to indicate that timeseries does
not need to include values for every day within the month,
set to TRUE to force this behavior.}
}
\value{
start_date first useable date in the timeseries (rounded to the first
        of the month)
}
\description{
Identifies the starting date of a monthly timeseries (forced to the 1st of
the month)
}
