% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_to_monthly_weather.R
\name{hourly_to_monthly_weather}
\alias{hourly_to_monthly_weather}
\title{Hourly to Monthly Weather}
\usage{
hourly_to_monthly_weather(filename, filedir = "data", skip_lines = 7)
}
\arguments{
\item{filename}{name of hourly weather file which includes:
\itemize{
  \item date (m/d/YYYY)
  \item time (H:MM)
  \item atmp - air temperature (deg C)
  \item relh - relative humidity (%)
  \item pcpn - precipitation (mm)
  \item rpet - reference potential evapotranspiration (mm/hr)
}}

\item{filedir}{directory in which hourly weather csv file resides}

\item{skip_lines}{number of rows in csv to skip before reading, default: 7}
}
\value{
monthly_weather, a data frame with the following columns:
\describe{
\item{month}{month of observation (m)}
\item{year}{year of observation (YYYY)}
\item{atmp_K}{mean monthly air temperature (K)}
\item{relh_pct}{mean monthly relative humidity (percent)}
\item{pcpn_mm}{monthly precipitaiton (mm)}
\item{rpet_mm}{monthly reference potential evapotranspiration (mm)}
}
}
\description{
This function takes hourly air temperature, relative humidity,
precipitaiton, and reference evapotranspiration as formatted by the Hancock,
 WI weather station and summarizes it at a monthly time step.
}
\references{
\url{https://enviroweather.msu.edu/weather.php?stn=hck}
}
