% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_gapfill.R
\name{iso_gapfill}
\alias{iso_gapfill}
\title{Fill Gaps in Stable Isotope Measurements}
\usage{
iso_gapfill(monthly_isotopes)
}
\arguments{
\item{monthly_isotopes}{a data frame with monthly isotope measurements for
all dates of desired timeseries (with NAs for months
with no measurements)}
}
\value{
monthly_isotopes - the same data frame provided to the function, but
with d18O_lake values filled in for all months
(except maybe the very first and last months).
}
\description{
Linearly interolates lake and GW values for dates in timeseries with missing
measurements. Depending on arguments, fills in precipitation values using 1)
measurments collected that month, and 2) Maribeth Kniffin's data, also from
Hancock station.
}
