% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_gw_site_names.R
\name{iso_gw_site_names}
\alias{iso_gw_site_names}
\title{Record Groundwater Sites Used for Isotopes}
\usage{
iso_gw_site_names(monthly_isotopes, isotopes)
}
\arguments{
\item{monthly_isotopes}{a data frame with monthly isotope measurements for
all dates of desired timeseries (with NAs for months
with no measurements)}

\item{isotopes}{a data frame with isotope measurements as formatted in the
\code{\link{isotopes}} dataset, subset to records for the lake
of interest.}
}
\value{
monthly_isotopes - same data frame with GWin_sites and GWout_sites
                           added as columns
}
\description{
Identifies which groundwater wells are included in monthly averages of
groundwater stable isotope measurements, then makes note of this in the
summary dataframe.
}
