% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lake_change.R
\name{iso_lake_change}
\alias{iso_lake_change}
\title{Estimates change in lake isotope measurement since previous month}
\usage{
iso_lake_change(monthly_isotopes)
}
\arguments{
\item{monthly_isotopes}{a data frame with monthly isotope measurements for
all dates of desired timeseries (with NAs for months
with no measurements)}
}
\value{
monthly_isotopes - the same data frame provided to the function, but
with delta_d18O_lake and delta_d2H_lake values
filled in for all months (except maybe the very
first).
}
\description{
Calculates change in d18O and d2H in the lake over the previous 30 days.
}
