% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_lake_gapfill.R
\name{iso_lake_gapfill}
\alias{iso_lake_gapfill}
\title{Fill Gaps in Lake Stable Isotope Measurements}
\usage{
iso_lake_gapfill(monthly_isotopes, static_lake = FALSE)
}
\arguments{
\item{monthly_isotopes}{a data frame with monthly isotope measurements for
all dates of desired timeseries (with NAs for months
with no measurements)}

\item{static_lake}{logical defaults to FALSE to use actual measurement for
each month. If TRUE, uses mean of fall (Sept-Nov) isotope
samples for the lake.}
}
\value{
monthly_isotopes - the same data frame provided to the function, but
                           with d18O_lake values filled in for all months
                           (except maybe the very first and last months).
}
\description{
Either uses the fall lake static isotope measurement for all periods of the
year (static_lake  = TRUE) or linearly interolates values for months with
missing measurements.
}
