% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_pcpn_gapfill.R
\name{iso_pcpn_gapfill}
\alias{iso_pcpn_gapfill}
\title{Fill Gaps in Precipitation Stable Isotope Measurements}
\usage{
iso_pcpn_gapfill(monthly_isotopes, extend_pcpn = TRUE,
  use_kniffin_pcpn = TRUE)
}
\arguments{
\item{monthly_isotopes}{a data frame with monthly isotope measurements for
all dates of desired timeseries (with NAs for months
with no measurements)}

\item{extend_pcpn}{logical defaults to TRUE to backfill months with no
precipitation}

\item{use_kniffin_pcpn}{logical defaults to TRUE to average in precipitation
isotope measurements by Maribeth Kniffin for each
month. Kniffin measurements taken at Hancock
Agricultural Research Station from May 2016 - Apr
2015 for each month.}
}
\value{
monthly_isotopes - the same data frame provided to the function, but
                           with d18O_lake values filled in for all months
                           (except maybe the very first and last months).
}
\description{
Fills in precipitation isotope measurements using 1) measurments collected in
subsequent month, assuming collector was active during month with missing
data, and 2) Maribeth Kniffin's data, also from Hancock station.
}
