% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{CSLS Isotope Measurements}
\format{A data frame with the following columns:
\describe{
  \item{date}{date of measurement}
  \item{lake}{lake associated with measurement}
  \item{site_id}{unique ID for site of measurement, e.g. "PRECIP", "LONG",
                 "LL-01"}
  \item{d18O}{isotopic composition for 18O (per mil)}
  \item{d2H}{isotopic composition for duterium (per mil)}
}}
\usage{
data(isotopes)
}
\description{
Measurements of 18O and 2H for the CSLS lakes (Plainfield Lake, Long Lake,
and Pleasant Lake).
}
\details{
Lake and groundwater isotope measurements are collected on site for each lake
while precipitation is measured using samples collected at the Hancock
Agricultural Research Station website
(\url{https://enviroweather.msu.edu/weather.php?stn=hck}) (station id: hck).
The Hancock Agricultural Research Station is located in Hancock, WI
(location: 44.1188, -89.533, elevation: 241m), approximately 8 miles from
Plainfield Lake, 8.5 miles from Long Lake, and 14.5 miles from Pleasant Lake.

Raw csv data is processed with the function
\code{\link{retrieve_csls_isotopes}} from the isoH2Obudget package. This
function subsets the raw data to only valid 18O and 2H measurements.
}
\seealso{
\code{\link{retrieve_csls_isotopes}}
}
\keyword{datasets}
