% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kniffin.R
\docType{data}
\name{kniffin}
\alias{kniffin}
\title{Kniffin Data for Long Lake}
\format{A list :
\describe{
  \item{isotopes}{monthly isotope measurements}
  \item{h2o_bal_inputs}{all inputs to the final water balance calculations,
                        including monthly weather and monthly isotope
                        measurements}
  \item{monthly_h2o_bal}{calculated monthly water balance}
}}
\usage{
data(kniffin)
}
\description{
Data from Maribeth Kniffin's isotope work at Long Lake in 2016-2017. Data
includes monthly weather, isotope measurments, and site name information.
}
\keyword{datasets}
