% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kniffin_isotopes.R
\docType{data}
\name{kniffin_isotopes}
\alias{kniffin_isotopes}
\title{Kniffin Isotope Measurements}
\format{A data frame with the following columns:
\describe{
  \item{date}{date (month) of measurement}
  \item{d18O_lake}{d18O isotopic composition, lake (per mil)}
  \item{d18O_pcpn}{d18O isotopic composition, precipitation (per mil)}
  \item{d18O_GWin}{d18O isotopic composition, inflowing groundwater (per mil)}
  \item{d18O_GWout}{d18O isotopic composition, inflowing groundwater (per
                    mil), not measured by Kniffin}
  \item{d2H_lake}{d2H isotopic composition, lake (per mil)}
  \item{d2H_pcpn}{d2H isotopic composition, precipitation (per mil)}
  \item{d2H_GWin}{d2H isotopic composition, inflowing groundwater (per mil)}
  \item{d2H_GWout}{d2H isotopic composition, inflowing groundwater (per
                    mil), not measured by Kniffin}
  \item{GWin_sites}{sites used for GWin measurement}
  \item{GWout_sites}{sites used for GWout measurement}
}}
\usage{
data(kniffin_isotopes)
}
\description{
Measurements of d18O taken by Maribeth Kniffin May 2016 - April 2017.
Includes d18O_lake and d18O_GWin taken at Long Lake and d18O_pcpn taken at
Hancock Agricultural Research Station.
}
\keyword{datasets}
