% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg_csv.R
\name{load_pkg_csv}
\alias{load_pkg_csv}
\title{Load csv from isoH2Obudget Installation}
\usage{
load_pkg_csv(filename, filedir = "system.file", skip_lines = 0)
}
\arguments{
\item{filename}{name of the csv file to load (e.g., "input_data.csv")}

\item{filedir}{name of file directory to search for this csv file. Defaults
to 'system.file' to search the inst/extdata location of the
isoH2Obudget package}

\item{skip_lines}{number of rows in csv file to skip before reading in.
Defaults to zero.}
}
\value{
df, a data frame created from the csv file
}
\description{
This function loads a csv file that by default resides in the installed
isoH2Obudget package (saved to inst/extdata during package creation). The
file location can be changed by providing a different file directory
argument.
}
