% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpn_iso_active.R
\docType{data}
\name{pcpn_iso_active}
\alias{pcpn_iso_active}
\title{Precipitation Collection Dates}
\format{A lubridate interval with time periods when precipitation collector
was actively deployed at Hancock Agricultural research station.}
\usage{
data(pcpn_iso_active)
}
\description{
Date intervals when precipitation collector at Hancock agricultural research
station was actively deployed for isotope measurements.
}
\keyword{datasets}
