% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_bal.R
\name{plot_annual_bal}
\alias{plot_annual_bal}
\title{Plot Balance}
\usage{
plot_annual_bal(annual_h2o_bal, as_vol = TRUE, as_pcnt = FALSE, text_size = 12)
}
\arguments{
\item{annual_h2o_bal}{a data frame with the date and all fluxes into and out
of the lake.}

\item{as_vol}{logical defaults to TRUE to plot water balance as a volume
(m^3). If FALSE, plots as depths.}

\item{as_pcnt}{logical defaults to FALSE to display as percents.}

\item{text_size}{size of font, defaults to 12 point}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function creates a plot object to evaluate the timeseries of d18O
measurements over time at each site.
}
