% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_functions.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot d18O}
\usage{
plot_balance(lake, monthly_h2o_bal)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{monthly_h2o_bal}{a data frame with the date and all fluxes into and out
of the lake.}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function creates a plot object to evaluate the timeseries of d18O
measurements over time at each site.
}
