% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_budget.R
\name{plot_budget}
\alias{plot_budget}
\title{Plot Balance}
\usage{
plot_budget(
  df,
  annual = FALSE,
  as_pcnt = FALSE,
  gw_only = FALSE,
  text_size = 12,
  title = "",
  fill_breaks = c("P", "E", "GWin", "GWout", "dV"),
  fill_labels = c("Precipitation", "Evaporation", "GW Inflow", "GW Outflow",
    expression(paste(Delta, " Lake Volume"))),
  fill_values = c("#1F78B4", "#A6CEE3", "#33A02C", "#B2DF8A", "#FB9A99")
)
}
\arguments{
\item{df}{a data frame with all fluxes into and out of a lake or lakes. Can
be either monthly values for one lake or annual values for multiple
lakes.}

\item{annual}{logical defaults to FALSE to create plot for monthly water
budgets. If TRUE, creates plot for annual water budgets.}

\item{as_pcnt}{logical defaults to FALSE to display as percents.}

\item{gw_only}{logical defaults to false to display all components of the
water budget.}

\item{text_size}{size of font, defaults to 12 point}

\item{title}{string to use for title, defaults to "".}

\item{fill_breaks}{vector with column names of variables to plot}

\item{fill_labels}{vector with display labels for plotted variables}

\item{fill_values}{vector with hex codes for colors associated with each
variable.}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function creates a plot object to evaluate the timeseries of d18O
measurements over time at each site.
}
