% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_functions.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot Colors}
\usage{
plot_colors(plot_obj, text_size = 12)
}
\arguments{
\item{plot_obj}{a plot object created with ggplot(data = actual_data, aes(x =
actual_x, y = actual_y, group = actual_group))}

\item{text_size}{size of font, defaults to 12 point}
}
\value{
new_obj - a plot object with aesthetics added
}
\description{
This function contains the all aesthetics for plotting the isotopic
signatures of all measured wells at a lake with all sites on the same plot,
denoted by color
}
