% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elev_area_vol.R
\name{plot_elev_area_vol}
\alias{plot_elev_area_vol}
\title{Plot lake elevation vs. area or volume for one or more lakes}
\usage{
plot_elev_area_vol(
  ydata,
  ylabel,
  yscales = scales::scientific,
  title = "",
  lakes = c("Pleasant", "Long", "Plainfield"),
  text_size = 10
)
}
\arguments{
\item{ydata}{colname of input data to display, defaults to "level_m" but can
also be "area_m2" or "vol_m3"}

\item{ylabel}{string to display for y-axis title, e.g., "Lake Elevation (m)"
or "expression(Lake~Volume~(m^{3}))"}

\item{yscales}{type of scaling for y-axis, defaults to scales::scientific
Other options include scales::number for decimal notation}

\item{title}{string to display for title, defaults to "".}

\item{lakes}{names of lakes to plot, defaults to include all:
"c("Pleasant", "Long", "Plainfield")"}

\item{text_size}{size of text, defaults to 10 pt.}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function creates a plot object that showcases the lake elevation vs.
lake area or lake elevation vs. lake volume relationship for the lakes of
interest.
}
