% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_functions.R
\name{plot_facet}
\alias{plot_facet}
\title{Plot Facet}
\usage{
plot_facet(plot_obj, df)
}
\arguments{
\item{plot_obj}{a plot object created with ggplot(data = actual_data, aes(x =
actual_x, y = actual_y, group = actual_group))}

\item{df}{a data frame with a column "date" representing the range of dates
plotted within plot_obj}
}
\value{
new_obj - a plot object with aesthetics added
}
\description{
This function contains the all aesthetics for plotting the isotopic
signatures of all measured wells at a lake with each site faceted out into
its own plot
}
