% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gradients.R
\name{plot_gradients}
\alias{plot_gradients}
\title{Plot water level gradients and isotope sample dates for wells}
\usage{
plot_gradients(
  df,
  date_intervals = "6 months",
  date_formats = "\%b '\%y",
  title = "",
  text_size = 10
)
}
\arguments{
\item{df}{data frame with the date, site_id, water level difference (diff_m),
and sample result, if exists (d18O).}

\item{date_intervals}{string with desired break interval for dates, defaults
to "6 months".}

\item{date_formats}{string with desired formatting for date labels, defaults
to "\%b '\%y" for "Oct '19"-type formatting.}

\item{title}{string to use for title, defaults to "". Could be
sprintf("\%s Lake -  Water Level Comparison", lake)}

\item{text_size}{size of font, defaults to 10 point}
}
\value{
plot_obj - a plot object with water level difference and dates of
isotope measurements for each lake site.
}
\description{
This function creates a plot object to compare the difference in lake level
and well level measurements for all sites at a given lake, paired with when
isotope measurements were collected at each site.
}
