% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input.R
\name{plot_input}
\alias{plot_input}
\title{Plot water balance inputs for one or more lakes}
\usage{
plot_input(
  df = NULL,
  ydata,
  ylabel,
  yscales = scales::scientific,
  geom_type = "col",
  title = "",
  facet_scales = "fixed",
  lakes = c("Pleasant", "Long", "Plainfield"),
  inputs_only = TRUE,
  date_intervals = "6 months",
  date_formats = "\%b '\%y",
  text_size = 10
)
}
\arguments{
\item{df}{optional data frame passed to function containing columns "date",
ydata, and "lake". Avoids need to run
\code{\link{runall_csls_inputs}} or
\code{\link{runall_csls_budget}} as part of this function.}

\item{ydata}{colname of input data to display, e.g., "dV_m3" or "d18O_evap"}

\item{ylabel}{string to display for y-axis title, e.g.,
"expression(Change~Lake~Volume~(m^{3}))" or
"expression(delta^18*'O')"}

\item{yscales}{type of scaling for y-axis, defaults to scales::scientific.
Other options include scales::number for decimal format,
scales::percent for percentages, etc.}

\item{geom_type}{defaults to "col" to dispay as bar plot. Set to "point" to
instead display as points, or "line" for lines.}

\item{title}{string to display for title, defaults to "".}

\item{facet_scales}{type of scaling of axis for facets, defaults to "fixed".
Can also be "free_y", "free_x", or "free".}

\item{lakes}{names of lakes to plot, defaults to include all:
"c("Pleasant", "Long", "Plainfield")"}

\item{inputs_only}{logical defaults to TRUE to trigger
\code{\link{runall_csls_inputs}}. Set to FALSE to trigger
\code{\link{runall_csls_budget}} instead (i.e., to obtain
groundwater inflow or outflow values).}

\item{date_intervals}{string with desired break interval for dates, defaults
to "6 months".}

\item{date_formats}{string with desired formatting for date labels, defaults
to "\%b '\%y" for "Oct '19"-type formatting.}

\item{text_size}{size of text, defaults to 10 pt.}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function creates a plot object that showcases a given input over time
for each lake of interest.
}
