% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_functions.R
\name{plot_iso}
\alias{plot_iso}
\title{Plot Isotope Relationships}
\usage{
plot_iso(lake, lake_isotopes)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{lake_isotopes}{a data frame with the date, lake, site_id, d18O, and d2H of
precip, groundwater, and lake measurements.}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function contains the all aesthetics for plotting the isotopic
signatures of all measured wells at a lake with each site faceted out into
its own plot
}
