% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lake_level.R
\name{plot_lake_level}
\alias{plot_lake_level}
\title{Plot lake levels for one or more lakes}
\usage{
plot_lake_level(
  ydata = "level_m",
  ylabel = "Lake Elevation (m)",
  yscales = scales::number,
  title = "",
  lakes = c("Pleasant", "Long", "Plainfield"),
  date_intervals = "6 months",
  date_formats = "\%b '\%y",
  text_size = 10
)
}
\arguments{
\item{ydata}{colname of input data to display, defaults to "level_m" but can
also be "area_m2" or "vol_m3"}

\item{ylabel}{string to display for y-axis title, e.g., "Lake Elevation (m)"
or "expression(Lake~Volume~(m^{3}))"}

\item{yscales}{type of scaling for y-axis, defaults to scales::number.
Other options include scales::scientific for scientific notation}

\item{title}{string to display for title, defaults to "".}

\item{lakes}{names of lakes to plot, defaults to include all:
"c("Pleasant", "Long", "Plainfield")"}

\item{date_intervals}{string with desired break interval for dates, defaults
to "6 months".}

\item{date_formats}{string with desired formatting for date labels, defaults
to "\%b '\%y" for "Oct '19"-type formatting.}

\item{text_size}{size of text, defaults to 10 pt.}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function creates a plot object that showcases lake levels (as levels,
areas, or volumes) over time for each lake of interest.
}
