% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_functions.R
\name{plot_levels}
\alias{plot_levels}
\title{Plot Levels}
\usage{
plot_levels(lake, water_level_diff, lake_isotopes, text_size = 12)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{water_level_diff}{a data frame with the date, site_no, obs_type,
level_m (of gw levels), and diff_m between daily gw level
observations and lake level observations.}

\item{lake_isotopes}{a data frame with the date, lake, site_id, d18O, and d2H of
precip, groundwater, and lake measurements.}

\item{text_size}{size of font, defaults to 12 point}
}
\value{
plot_obj - a plot object with water level difference and dates of
                   isotope measurements for each lake site.
}
\description{
This function creates a plot object to compare the difference in lake level
and well level measurements for all sites at a given lake, paired with when
isotope measurements were collected and what the d18O results were.
}
