% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LMWL_all.R
\name{plot_LMWL_all}
\alias{plot_LMWL_all}
\title{Plot isotope relationships}
\usage{
plot_LMWL_all(
  lakes = c("Pleasant", "Long", "Plainfield"),
  site_types = c("precipitation", "upgradient", "downgradient", "deep", "lake"),
  facets_on = TRUE,
  upgradient_off = FALSE,
  downgradient_off = FALSE,
  deep_off = FALSE,
  text_size = 12
)
}
\arguments{
\item{lakes}{lakes of interest, defaults to
c("Pleasant", "Long", "Plainfield")}

\item{site_types}{type of sites, defaults to c("precipitation", "upgradient",
"downgradient", "deep", "lake")}

\item{facets_on}{defaults to TRUE to facet results by lake}

\item{upgradient_off}{logical defaults to FALSE to include upgradient wells}

\item{downgradient_off}{logical defaults to FALSE to include downgradient wells}

\item{deep_off}{logical defaults to FALSE to include deep wells}

\item{text_size}{size of font, defaults to 12 point}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function contains the all aesthetics for plotting the isotopic
signatures of precipitation, groundwater, and lake water based on all
samples. Also displays the Local Meteoric Water Line and Local Evaporation
Water Line. Options to facet by lake (default) or plot all on one graph. Can
turn of groundwater samples individually if desired.
}
