% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LMWL_sites.R
\name{plot_LMWL_sites}
\alias{plot_LMWL_sites}
\title{Plot isotope relationships by site}
\usage{
plot_LMWL_sites(lake, text_size = 12)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{text_size}{size of font, defaults to 12 point}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function contains the all aesthetics for plotting the isotopic
signatures of all measured wells at a lake with each site faceted out into
its own plot
}
