% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_csls_lst.R
\name{retrieve_csls_lst}
\alias{retrieve_csls_lst}
\title{Retrieve Central Sands Lake Surface Temperature}
\usage{
retrieve_csls_lst(filename, filedir = "system.file")
}
\arguments{
\item{filename}{name of SWIMS csv file which includes the following fields:
\itemize{
  \item DNR Parameter - unique DNR number for parameter. Assume lake
        surface temperatures are stored as "10" for "Temperature Field"
  \item Result - numeric value of measurement
  \item Units - units of result
  \item Result Depth - depth at which measurement taken, asume is "0
        Meters" for lake surface temperatures
  \item WBIC - Wisconsin Water Body Identification Code
}}

\item{filedir}{directory in which weather csv file resides, defaults to
'system.file' to instruct it to look within installed package
files.}
}
\value{
lst, a data frame with the following columns:
\describe{
\item{WBIC}{Water Body Identification Code for lake}
\item{date}{date and time of observation}
\item{ltmp}{lake surface temperature (degrees C)}
\item{units}{units of lake surface temperature, should all be "DEGREES C"}
}
}
\description{
This function retrieves every-other-week measurements of lake surface
temperatures associated with CSLS from SWIMS. Extracts only DNR Parameter 10
with Result Depth of "0 Meters".
}
\references{
\url{https://dnr.wi.gov/topic/surfacewater/swims/}
}
