% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_csls_stage_volume.R
\name{retrieve_csls_stage_volume}
\alias{retrieve_csls_stage_volume}
\title{Retrieve Central Sands Bathymetry Information}
\usage{
retrieve_csls_stage_volume(filename, filedir = "system.file")
}
\arguments{
\item{filename}{name of stage volume csv exported from ArcGIS}

\item{filedir}{name of directory with both csv files, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}
}
\value{
stage_vol, a data frame with the following columns:
\describe{
\item{lake}{lake associated with measurement}
\item{stage_m}{stage of lake (m) above mean sea level}
\item{surf_area_m2}{surface area of lake (m^2) at given elevation}
\item{volume_m3}{volume of lake (m^3) at given elevation}
}
}
\description{
Imports static bathymetry data for the CSLS lakes which was exported from
ArcGIS. Retains only stage (m), surface area (m^2), and volume (m^3).
}
