% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_csls_water_levels.R
\name{retrieve_csls_water_levels}
\alias{retrieve_csls_water_levels}
\title{Retrieve Daily Central Sands Water Levels}
\usage{
retrieve_csls_water_levels()
}
\value{
water_levels, a data frame with the following columns
\describe{
\item{date}{date of water level measurement}
\item{site_no}{UGSG site number}
\item{obs_type}{type of observation (LK = lake level, GW = groundwater
                level)}
\item{level_m}{water level in meters above mean sea level}
}
}
\description{
Downloads current feature layer from DNR feature services with water quantity
measurements in the Central Sands (summarized at a daily time step). Feature
layer of interest is DG_HiCap > DG_CSLS_QUANT_MON_WTM_EXT > Layer 2
(W13101.WU_CSLS_QUANT_DATA).
}
