% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_csls_weather.R
\name{retrieve_csls_weather}
\alias{retrieve_csls_weather}
\title{Retrieve Hourly Central Sands Weather}
\usage{
retrieve_csls_weather(filename, filedir = "system.file",
  skip_lines = 7)
}
\arguments{
\item{filename}{name of Hancock weather csv}

\item{filedir}{directory in which weather csv file resides, defaults to
'system.file' to instruct it to look within installed package
files.}

\item{skip_lines}{[numeric] number of rows in csv to skip before reading in
data, defaults to 7.}
}
\value{
weather, a data frame with the following columns for hourly weather:
\describe{
\item{date}{date and time of weather observation}
\item{atmp}{air temperature (deg C)}
\item{P}{precipitation (mm)}
\item{RH}{relative humidity (percent)}
\item{ET}{reference potential evapotranspiration (mm)}
}
}
\description{
This function retrieves hourly weather from the Hancock, WI weather station.
Required weather fields include air temperature (deg C), relative humidity
(%), precipitation (mm), and reference evapotranspiration (mm). NA values
filled in via linear interpolation.
}
