% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runall_isoH2Obudget.R
\name{runall_isoH2Obudget}
\alias{runall_isoH2Obudget}
\title{Run isoH2Obudget Start to Finish}
\usage{
runall_isoH2Obudget(lake, dataset = "csls")
}
\arguments{
\item{lake}{name of lake (e.g., Pleasant, Long, or Plainfield)}

\item{dataset}{dataset type to process. Expects "csls" or "kniffin".}
}
\value{
monthly_h2o_bal, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{P}{monthly precipitation (mm)}
\item{E}{monthly evapotranspiration (mm)}
\item{GWin}{monthly groundwater inflow to the lake (mm)}
\item{GWout}{monthly groundwater outflow to the lake (mm)}
\item{dV}{monthly change in lake volume (mm)}
}
}
\description{
Runs all major functions in the isoH2Obudget package, from loading in raw
data to the final monthly water balance
}
