% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_n_months.R
\name{start_n_months}
\alias{start_n_months}
\title{Monthly Timeseries Start and Number}
\usage{
start_n_months(date_vector, all_days = TRUE)
}
\arguments{
\item{date_vector}{a vector of class "Date"}

\item{all_days}{logical. If TRUE, timeseries start and end must include all
days within the month (default = TRUE)}
}
\value{
month_info, a list
\describe{
\item{start_date}{Start month of timeseries (YYYY-MM-01)}
\item{nmonths}{number of months in timeseries}
}
}
\description{
This function takes a date vector and determines the start month, start
year, and number of months in a time series. It includes an option to
require a complete daily timeseries in order to include a month.
}
