% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_lake_functions.R
\name{subset_gw_levels}
\alias{subset_gw_levels}
\title{Subset Groundwater Levels}
\usage{
subset_gw_levels(lake, water_levels, site_dictionary)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{water_levels}{a data frame with the date, site_no, obs_type, and
level_m of daily water level observations, as in the
\code{\link{water_levels}} dataset.}

\item{site_dictionary}{a data frame with the lake, obs_type, site_id,
SWIMS_station, USGS_id, WBIC, and static_iso_type of
measurement sites, as in the
\code{\link{site_dictionary}} dataset.}
}
\value{
gw_levels - a data frame with the same columns as in water_levels
                    (see the \code{\link{water_levels}} dataset), but subset
                    for only groundwater wells at the lake of interest. The
                    site_id (e.g., LL-01) is also attached.
}
\description{
Subsets larger water level data frame by USGS_id to retrieve only groundwater
records for a given lake.
}
\seealso{
\code{\link{retrieve_csls_water_levels}},
         \code{\link{water_levels}}, \code{\link{site_dictionary}},
         \code{vignette("retrieve_csls_input_data", package = "isoH2Obudget")}
}
