% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_isotopes.R
\name{subset_isotopes}
\alias{subset_isotopes}
\title{Subset Isotopes}
\usage{
subset_isotopes(lake, isotopes)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{isotopes}{a data frame with the date, lake, site_id, d18O, and d2H of
isotope measurements, as in the \code{\link{isotopes}}
dataset}
}
\value{
isotopes - a data frame with the same columns as in isotopes (see the
                   \code{\link{isotopes}} dataset), but subset for only
                   precipitation measurements plus lake and groundwater
                   measurements at the lake of interest.
}
\description{
Subsets larger isotopes data frame to retain only precipitation measurements
plus lake and groundwater measurements associated with the lake of interest.
}
\seealso{
\code{\link{retrieve_csls_isotopes}},
         \code{\link{isotopes}},
         \code{vignette("retrieve_csls_input_data", package = "isoH2Obudget")}
}
