% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_lake_functions.R
\name{subset_lst}
\alias{subset_lst}
\title{Subset Lake Surface Temperature}
\usage{
subset_lst(lake, lst, site_dictionary)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{lst}{a data frame with the WBIC, date, ltmp, and units of
every-other-week lake surface temperature measurements, as in the
\code{\link{lst}} dataset}

\item{site_dictionary}{a data frame with the lake, obs_type, site_id,
SWIMS_station, USGS_id, WBIC, and static_iso_type of
measurement sites, as in the
\code{\link{site_dictionary}} dataset.}
}
\value{
lst - a data frame with the same columns as in lst (see the
              \code{\link{lst}} dataset), but subset for only lake surface
              temperatures atthe lake of interest.
}
\description{
Subsets larger lake surface temperature data frame to retain only those
measurements associated with the lake of interest.
}
\seealso{
\code{\link{retrieve_csls_lst}},
         \code{\link{lst}}, \code{\link{site_dictionary}},
         \code{vignette("retrieve_csls_input_data", package = "isoH2Obudget")}
}
