% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_site_dictionary.R
\name{subset_site_dictionary}
\alias{subset_site_dictionary}
\title{Subset Site Dictionary}
\usage{
subset_site_dictionary(lake, site_dictionary)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{site_dictionary}{a data frame with the lake, obs_type, site_id,
SWIMS_station, USGS_id, WBIC, and static_iso_type of
measurement sites, as in the
\code{\link{site_dictionary}} dataset}
}
\value{
site_dictionary - a data frame with the same columns as in
                          site_dictionary (see the
                          \code{\link{site_dictionary}} dataset), but subset
                          for only those records for the lake of interest.
}
\description{
Subsets larger site_dictionary data frame to retain only those records
related to the lake of interest.
}
\seealso{
\code{\link{retrieve_csls_site_dictionary}},
         \code{\link{site_dictionary}},
         \code{vignette("retrieve_csls_input_data", package = "isoH2Obudget")}
}
