% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_lake_functions.R
\name{subset_stage_vol}
\alias{subset_stage_vol}
\title{Subset Stage Volume Relationship}
\usage{
subset_stage_vol(lake, stage_vol)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{stage_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the \code{\link{stage_vol}} dataset.}
}
\value{
stage_vol - a data frame with the same columns as in stage_vol (see the
                   \code{\link{stage_vol}} dataset), but subset for only
                   those records for the lake of interest.
}
\description{
Subsets larger stage_vol data frame to retain only those records related to
the lake of interest.
}
\seealso{
\code{\link{retrieve_csls_stage_volume}},
         \code{\link{stage_vol}},
         \code{vignette("retrieve_csls_input_data", package = "isoH2Obudget")}
}
