% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_d18O_evap.R
\name{summarise_d18O_evap}
\alias{summarise_d18O_evap}
\title{Summarize Evaporation d18O}
\usage{
summarise_d18O_evap(monthly_weather, monthly_lst, monthly_isotopes)
}
\arguments{
\item{monthly_weather}{a data frame with monthly weather including date,
atmp_K (air temperature, K), RH_pct (relative humidity,
percent), P_mm (precipitation, mm), and ET_mm
(evapotranspiration, mm) as output by
\code{\link{summarise_weather}}}

\item{monthly_lst}{a data frame as output by \code{\link{summarise_lst}}}

\item{monthly_isotopes}{a data frame as output by
\code{\link{summarise_isotopes}}}
}
\value{
monthly_isotopes with d18O_evap column added
}
\description{
Calculates Evaporation d18O based on monthly weather, lake surface
temperature, and isotope measurements.
}
