% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dV.R
\name{summarise_dV}
\alias{summarise_dV}
\title{Monthly Change in Lake Volume}
\usage{
summarise_dV(lake_levels, timeseries)
}
\arguments{
\item{lake_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to lake level records for the lake of interest.}

\item{timeseries}{a vector of all months in the common timeseries among input
datasets}
}
\value{
monthly_dV, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{dV}{change in lake volume during the month as a depth (mm)}
}
}
\description{
Calculates the monthly change in lake volume based on daily lake levels and a
stage-volume relationship from bathymetry analysis. .
}
