% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_inputs.R
\name{summarise_inputs}
\alias{summarise_inputs}
\title{Summarize Inputs over Common Timeseries}
\usage{
summarise_inputs(weather, lst, isotopes, lake_levels, gw_levels, stage_vol,
  site_dictionary, static_gw = FALSE, median_threshold = 0.01,
  static_lake = FALSE, use_kniffin_pcpn = TRUE,
  pcpnfile = "csls_isotope_precipitation_deployment.csv",
  pcpndir = "system.file")
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including date, atmp (air
temperature, deg C), RH (relative humidity, percent), P
(precipitation, mm), Rs (incoming solar radiation, MJ/m^2),
and wind (wind speed, m/s) as formatted in the
\code{\link{weather}} dataset}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the \code{\link{lst}} dataset, subset
for a single lake.}

\item{isotopes}{a data frame with isotopes measurements as formatted in the
\code{\link{isotopes}} dataset, subset for a single lake.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to groundwater level records at the lake of
interest.}

\item{stage_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the \code{\link{stage_vol}} dataset, subset
for a single lake.}

\item{site_dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
\code{\link{site_dictionary}} dataset, subset to
records for the lake of interest.}

\item{static_gw}{logical defaults to FALSE to use lake_levels and gw_levels
to define upgradient/downgradient wells at each measurement
date. If TRUE, uses static definitions of
upgradient/downgradient wells in site dictionary.}

\item{median_threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in
order to classify groundwater measurement.}

\item{static_lake}{logical defaults to FALSE to use actual measurement for
each month. If TRUE, uses mean of fall (Sept-Nov) isotope
samples for the lake.}

\item{use_kniffin_pcpn}{logical defaults to TRUE to average in precipitation
measurements by Maribeth Kniffin for each month.}

\item{pcpnfile}{name of file with dates of precipitation collector deployment.
Defaults to "csls_isotope_precipitation_deployment.csv"}

\item{pcpndir}{name of directory with file with dates of precipitation
collector deployment. Defaults to "system.file" to indicate is
stored within inst/extdata within the installed isoH2Obudget
package files.}
}
\value{
h2o_bal_inputs, a data frame ordered by date with data only for the
        overlap timeseries for:
\describe{date}{dates of monthly timeseries (1st of the month) for months with
                complete overlapping input data}
\describe{atmp_degC}{mean air temperature (degrees C)}
\describe{RH_pct}{mean relative humidity (percent)}
\describe{P_mm}{total precipitation (mm)}
\describe{ET_mm}{total lake evaporation (mm)}
\describe{atmp_K}{mean air temperature (Kelvin)}
\describe{ltmp_K}{mean lake surface temperature (degrees C)}
\describe{dV}{change in lake volume (mm)}
\describe{d18O_lake}{mean stable isotope measurement for lake}
\describe{d18O_pcpn}{mean stable isotope measurement for precipitation}
\describe{d18O_GWin}{mean stable isotope measurement for inflowing
                     groundwater}
\describe{d18O_GWout}{mean stable isotope measurement for outflowing
                      groundwater}
\describe{d18O_evap}{mean stable isotope measurement for evaporation}
\describe{GWin_sites}{wells included in mean inflowing groundwater stable
                      isotope measurement}
\describe{GWout_sites}{wells included in mean outflowing groundwater stable
                      isotope measurement}
}
\description{
This function takes raw weather, lake surface temperature, isotope, lake
level, groundwater level, stage-volume inputs and summarises water balance
inputs over a common timeseries.
}
