% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_isotopes.R
\name{summarise_isotopes}
\alias{summarise_isotopes}
\title{Monthly Isotope Measurements}
\usage{
summarise_isotopes(isotopes, dictionary, analysis, static_gw = FALSE,
  lake_levels = NULL, gw_levels = NULL, threshold = 0.01,
  static_lake = FALSE, use_kniffin_pcpn = TRUE, extend_pcpn = TRUE)
}
\arguments{
\item{isotopes}{a data frame with isotopes measurements as formatted in the
isotopes dataset, subset for a single lake.}

\item{dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
dictionary dataset, subset to
records for the lake of interest.}

\item{analysis}{a list with dates (POSIXct) to analyze, and intervals
(lubridate interval) covering the month before each analysis
date.}

\item{static_gw}{logical defaults to FALSE to use lake_levels and gw_levels
to define upgradient/downgradient wells at each measurement
date. If TRUE, uses static definitions of
upgradient/downgradient wells in site dictionary.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to groundwater level records at the lake of
interest.}

\item{threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in order
to classify groundwater measurement.}

\item{static_lake}{logical defaults to FALSE to use actual measurement for
each month. If TRUE, uses mean of fall (Sept-Nov) isotope
samples for the lake.}

\item{use_kniffin_pcpn}{logical defaults to TRUE to average in precipitation
isotope measurements by Maribeth Kniffin for each
month. Kniffin measurements taken at Hancock
Agricultural Research Station from May 2016 - Apr
2015 for each month.}

\item{extend_pcpn}{logical defaults to TRUE to backfill months with no
precipitation}
}
\value{
monthly_isotopes, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{d18O_pcpn}{mean precipitation stable isotope measurement for the month}
\item{d18O_lake}{mean lake stable isotope measurement for the month}
\item{d18O_GWin}{mean groundwater inflow stable isotope measurement for the
                 month}
\item{d18O_GWout}{mean groundwater outflow stable isotope measurement for the
                 month}
\item{GWin_sites}{well sites used in mean d18O_GWin each month}
\item{GWout_sites}{well sites used in mean d18O_GWout each month}
}
}
\description{
Summarizes isotope measurements at a monthly timestep for a single lake
}
