% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_lst.R
\name{summarise_lst}
\alias{summarise_lst}
\title{Monthly Lake Surface Temperature}
\usage{
summarise_lst(lst, timeseries)
}
\arguments{
\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the lst dataset, subset
for a single lake.}

\item{timeseries}{a vector of all months in the common timeseries among input
datasets}
}
\value{
monthly_lst, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{ltmp_K}{mean monthly lake surface temperature (degrees K)}
}
}
\description{
Summarizes sub-monthly lake surface temperature at a monthly timestep for a
single lake
}
