% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_weather.R
\name{summarise_weather}
\alias{summarise_weather}
\title{Monthly Weather}
\usage{
summarise_weather(weather, lst, stage_vol, site_dictionary, timeseries,
  wind_elev = 3, z0 = 0.02, Lz = 90, lake_albedo = 0.08)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including date, atmp (air
temperature, deg C), RH (relative humidity, percent), P
(precipitation, mm), Rs (incoming solar radiation, MJ/m^2),
and wind (wind speed, m/s) as formatted in the
weather dataset.}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the lst dataset, subset
for a single lake.}

\item{stage_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the stage_vol dataset, subset
for a single lake.}

\item{site_dictionary}{a data frame with the lake, obs_type, site_id,
SWIMS_station, USGS_id, WBIC, and static_iso_type of
measurement sites, as in the
site_dictionary dataset, subset for a
single lake.}

\item{timeseries}{a vector of all months in the common timeseries among input
datasets}

\item{wind_elev}{height of wind measurement, defaults to 3m for Hancock station.}

\item{z0}{aeorodynamic roughness of landcover at site of weather measurements
(m), defaults to 0.02 for a grass.}

\item{Lz}{the longitude of the local timezone (degrees west of Greenwich,
ranges from 0 to 360 degrees). Defaults to 90 for Central Time
Zone, USA.}

\item{lake_albedo}{albedo of the lake, defaults to 0.08 for open water.}
}
\value{
monthly_weather, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{atmp_K}{mean monthly air temperature (K)}
\item{RH_pct}{mean monthly relative humidity (percent)}
\item{P_mm}{monthly precipitaiton (mm)}
\item{ET_mm}{monthly lake evaporation (mm)}
}
}
\description{
Summarizes sub-monthly weather at a monthly timestep
}
