% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_levels.R
\docType{data}
\name{water_levels}
\alias{water_levels}
\title{CSLS Daily Lake and Groundwater Levels}
\format{A data frame with the following columns:
\describe{
\item{date}{date of water level measurement}
\item{site_no}{UGSG site number}
\item{obs_type}{type of observation (LK = lake level, GW = groundwater
                level)}
\item{level_m}{water level in meters above mean sea level}
}}
\source{
\url{https://uadnrmaps.wi.gov/arcgis/rest/services/DG_HiCap/DG_CSLS_QUANT_MON_WTM_EXT/FeatureServer/2}
}
\usage{
data(water_levels)
}
\description{
Daily lake level and groundwater level observations for the CSLS lakes and
related groundwater monitoring wells.
}
\details{
Raw data are retrieved from DNR Water Use section ArcGIS feature services for
the Central Sands. Water quantity information is summarized at a daily time
step for these measurements. The feature layer of interest is
\url{https://uadnrmaps.wi.gov/arcgis/rest/services/DG_HiCap/DG_CSLS_QUANT_MON_WTM_EXT/FeatureServer/2}.

Raw csv data is processed with the function
\code{\link{retrieve_csls_water_levels}} from the isoH2Obudget package. This
function retains only lake levels and groundwater levels from the larger set
of groundwater quantity measurements in the Central Sands.
}
\seealso{
\code{\link{retrieve_csls_water_levels}}
}
\keyword{datasets}
