% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_rpet.R
\docType{data}
\name{weather_rpet}
\alias{weather_rpet}
\title{CSLS Hourly Weather Data - RPET version}
\format{A data frame with hourly records for:
\describe{
\item{date}{date and time of weather observation}
\item{atmp}{air temperature (deg C)}
\item{P}{precipitation (mm)}
\item{RH}{relative humidity (percent)}
\item{ET}{reference potential evapotranspiration (mm)}
}}
\source{
\url{https://enviroweather.msu.edu/weather.php?stn=hck}
}
\usage{
data(weather)
}
\description{
Hourly weather data from the Hancock Agricultural Research Station website
(\url{https://enviroweather.msu.edu/weather.php?stn=hck}) (station id: hck)
}
\details{
The Hancock Agricultural Research Station is located in Hancock, WI
(location: 44.1188, -89.533, elevation: 241m), approximately 8 miles from
Plainfield Lake, 8.5 miles from Long Lake, and 14.5 miles from Pleasant Lake.

To download raw data, go to the Hancock Station website and click on "Weather
Station at Hancock" or "More weather for this station" to be routed to a page
with options for "Custom Reports (Data-on-Demand)". Select Data Type "Hourly
Data", then check Air Temperature, Precipitation, Relative Humidity, and
Reference Potential Evapotranspiration. Select starting date and end date,
display units (metric), and output (csv), then "Generate Report". Note that
the earliest starting date allowed (as of January and September 2019) is
February 15, 2018.

Raw csv data is processed with the function
\code{\link{retrieve_csls_weather}} from the isoH2Obudget package. This
function linearly interpolates any timestamps with missing values in the raw
data.
}
\seealso{
\code{\link{retrieve_csls_weather}}
}
\keyword{datasets}
