% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GW_functions.R
\name{calculate_GW_outflow}
\alias{calculate_GW_outflow}
\title{Groundwater Inflow}
\usage{
calculate_GW_outflow(P, E, GWin, dVdt = 0)
}
\arguments{
\item{P}{precipitation (mm)}

\item{E}{evapotranspiration (mm)}

\item{GWin}{groundwater inflow (mm)}

\item{dVdt}{change in lake volume storage (mm)}
}
\value{
GWout - the groundwtaer outflow from the lake (mm)
}
\description{
Calculates the groundwater inflow to a lake based on equation 4 in
Krabbenhoft et al. (1990).
}
\references{
Krabbenhoft, D. P., C. J. Bowser, M. P. Anderson, and J. W.
Valley. (1990). Estimating Groundwater Exchange with Lakes: 1. The Stable
Isotope Mass Balance Method. Water Resources Research, 26(10):2445-2453.
https://doi.org/10.1029/WR026i010p02445
}
