% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lake_evap.R
\name{calculate_lake_evap}
\alias{calculate_lake_evap}
\title{Calculate Lake Evaporation}
\usage{
calculate_lake_evap(
  weather,
  lst,
  elev_area_vol,
  dictionary,
  Lz = 90,
  wind_elev = 3,
  z0 = 0.02,
  no_condensation = FALSE
)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including including date
(POSIXct), atmp (air temperature, deg C), P (precipitation, mm), RH
(relative humidity, percent), Rs (solar radiation, MJ), and wind (wind
speed, m/s)}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the lst_HOBO dataset, subset
for a single lake.}

\item{elev_area_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the elev_area_voldataset, subset
for a single lake.}

\item{dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
dictionary dataset, subset to
records for the lake of interest.}

\item{Lz}{the longitude of the local timezone (degrees west of Greenwich,
ranges from 0 to 360 degrees). Defaults to 90 for Central Time
Zone, USA.}

\item{wind_elev}{height of wind measurement, defaults to 3m for Hancock
station.}

\item{z0}{aeorodynamic roughness of landcover at site of weather measurements
(m), defaults to 0.02 for a grass.}

\item{no_condensation}{logical defaults to FALSE to include days with
condensation in daily lake evaporation calculations.
If TRUE, sets all days with condensation (i.e.,
negative lake evaporation) to zero.}
}
\value{
weather, a list with daily weather information including:
\describe{
\item{date}{day of each weather observation}
\item{atmp_min}{minimum air temperature for the day (deg C)}
\item{atmp_max}{maximum air temperature for the day (deg C)}
\item{RH_min}{minimum relative humidity for the day (percent)}
\item{RH_max}{maximum relative humidity for the day (percent)}
\item{P}{total precipitation for the day (mm)}
\item{ET}{total lake evaporation for the day (mm)}
}
}
\description{
Calculates Lake Evaporation
}
