% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_iso_site_dynamic.R
\name{classify_iso_site_dynamic}
\alias{classify_iso_site_dynamic}
\title{Classify Stable Isotope Measurements (Dynamic)}
\usage{
classify_iso_site_dynamic(isotopes, water_levels, lake,
  site_file = "csls_site_dictionary.csv", filedir = "system.file",
  Xday = 7)
}
\arguments{
\item{isotopes}{a data frame with the following columns
\itemize{
\item date - date and time of sample collection (Date)
\item site_id - unique site id for where measurement was taken (e.g.,
                LL-01)
\item d18O - stable isotope measurement for d18O at this site
\item d2H - stable isotope measurement for for duterium at this site
}}

\item{water_levels}{a data frame with the following columns
\itemize{
\item date - date of water level measurement
\item site_no - UGSG site number
\item obs_type - type of observation (LK = lake level, GW = groundwater
                 level)
\item level_m - water level in meters above mean sea level
}}

\item{lake}{name of lake to analyze (e.g., Pleasant, Long, or Plainfield).
This name corresponds to lake names in the site dictionary "lake"
column.}

\item{site_file}{filname of site dictionary with the following columns
\itemize{
\item lake - name of lake associated with the measurement site. Corresponds
             to lake name argument in this function.
\item obs_type - type of observation (LK = lake, GW = groundwater
                 monitoring well)
\item site_id - unique site id for measurement site (e.g., LL-01).
                Corresponds to site_id in "isotopes" data frame.
\item SWIMS_station_id - SWIMS station id, if exists for this site.
\item USGS_id - USGS site number. Corresponds to site_no in "water_levels"
                data frame.
\item WBIC - water body identification code (WBIC), for lake sites only.
}}

\item{filedir}{name of directory with both csv files, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}

\item{Xday}{number of days to consider when calculating average water level
before isotope measurements, defaults to 7 days.}
}
\value{
isotopes - the same data frame provided to the function, but with an
                   additional column for "site_type" with values of
                   "precipitation", "lake", "upgradient", or "downgradient".
}
\description{
Classifies the site type of isotope measurements as precipitation, lake,
upgradient, or downgradient. Upgradient vs. downgradient wells are determined
by comparing the 7-day mean lake level and well level leading up to isotope
measurements.
}
