% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_iso_site_functions.R
\name{classify_iso_site_dynamic}
\alias{classify_iso_site_dynamic}
\title{Classify Stable Isotope Measurements (Dynamic)}
\usage{
classify_iso_site_dynamic(isotopes, lake_levels, gw_levels,
  site_dictionary, Xday = 7)
}
\arguments{
\item{isotopes}{a data frame with isotope measurements as formatted in the
\code{\link{isotopes}} dataset, subset to records for the lake
of interest.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to groundwater level records at the lake of
interest.}

\item{site_dictionary}{a data frame with site id numbers as formatted in the
\code{\link{site_dictionary}} dataset, subset to
records for the lake of interest.}

\item{Xday}{number of days to consider when calculating average water level
before isotope measurements, defaults to 7 days.}
}
\value{
isotopes - the same data frame provided to the function, but with an
                   additional column for "site_type" with values of
                   "precipitation", "lake", "upgradient", or "downgradient".
}
\description{
Classifies the site type of isotope measurements as precipitation, lake,
upgradient, or downgradient. Upgradient vs. downgradient wells are determined
by comparing the 7-day mean lake level and well level leading up to isotope
measurements.
}
