% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_iso_site_static.R
\name{classify_iso_site_static}
\alias{classify_iso_site_static}
\title{Classify Stable Isotope Measurements (Static)}
\usage{
classify_iso_site_static(isotopes, lake,
  site_file = "csls_site_dictionary.csv", filedir = "system.file")
}
\arguments{
\item{isotopes}{a data frame with the following columns
\itemize{
\item date - date and time of sample collection (Date)
\item site_id - unique site id for where measurement was taken (e.g.,
                LL-01)
\item d18O - stable isotope measurement for d18O at this site
\item d2H - stable isotope measurement for for duterium at this site
}}

\item{lake}{name of lake to analyze (e.g., Pleasant, Long, or Plainfield).
This name corresponds to lake names in the site dictionary "lake"
column.}

\item{site_file}{filname of site dictionary with the following columns
\itemize{
\item lake - name of lake associated with the measurement site. Corresponds
             to lake name argument in this function.
\item obs_type - type of observation (LK = lake, GW = groundwater
                 monitoring well)
\item site_id - unique site id for measurement site (e.g., LL-01).
                Corresponds to site_id in "isotopes" data frame.
\item SWIMS_station_id - SWIMS station id, if exists for this site.
\item USGS_id - USGS site number. Corresponds to site_no in "water_levels"
                data frame.
\item WBIC - water body identification code (WBIC), for lake sites only.
}}

\item{filedir}{name of directory with both csv files, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}
}
\value{
isotopes - the same data frame provided to the function, but with an
                   additional column for "site_type" with values of
                   "precipitation", "lake", "upgradient", "downgradient", or
                   NA.
}
\description{
Classifies the site type of isotope measurements as precipitation, lake,
upgradient, or downgradient. Upgradient vs. downgradient wells are determined
by static definitions in the site dictionary.
}
