% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap_d18O_atm}
\alias{d18O_evap_d18O_atm}
\title{Atmosphere d18O}
\usage{
d18O_evap_d18O_atm(d18O_pcpn, alpha)
}
\arguments{
\item{d18O_pcpn}{isotopic composition of precipitation}

\item{alpha}{equilibrium isotope fractionation factor at the temperature of
the air-water interface}
}
\value{
d18O_atm - the d18O isotope composition of the atmosphere
}
\description{
Calculates the d18O isotope composition of the atmosphere
}
