% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap_isotope_frac}
\alias{d18O_evap_isotope_frac}
\title{Equilibrium Isotope Fractionation Factor}
\usage{
d18O_evap_isotope_frac(ltmp)
}
\arguments{
\item{ltmp}{lake surface temperature (K)}
}
\value{
alpha - the equilibrium isotope fractionation factor
}
\description{
Calculates the equilibrium isotope fractionation factor at the temperature of
the air-water interface.
}
