% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap_kinetic_frac}
\alias{d18O_evap_kinetic_frac}
\title{Kinetic Fractionation Factor}
\usage{
d18O_evap_kinetic_frac(h, K = 14.3)
}
\arguments{
\item{h}{relative humidity normalized to the temperature of the surface water}

\item{K}{constant determinded by wind tunnel experiments for different
isotopes, defaults to K(18O) = 14.3.}
}
\value{
delta_epsilon - the kinetic fractionation factor
}
\description{
Calculates the kinetic fractionation factor
}
