% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O_evap_functions.R
\name{d18O_evap_total_frac}
\alias{d18O_evap_total_frac}
\title{Total Fractionation Factor}
\usage{
d18O_evap_total_frac(alpha, delta_epsilon)
}
\arguments{
\item{alpha}{equilibrium isotope fractionation factor at the temperature of
the air-water interface (-)}

\item{delta_epsilon}{kinetic fractionation factor (-)}
}
\value{
epsilon - the total fractionation factor (-)
}
\description{
Calculates the total fractionation factor based on the definition of epsilon
for equation 5 in Krabbenhoft et al. (1990). Note that alpha* is equivalent
to 1/alpha.
}
\references{
Krabbenhoft, D. P., C. J. Bowser, M. P. Anderson, and J. W.
  Valley. (1990). Estimating Groundwater Exchange with Lakes: 1. The Stable
  Isotope Mass Balance Method. Water Resources Research, 26(10):2445-2453.
  https://doi.org/10.1029/WR026i010p02445
}
