% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_functions.R
\name{find_timeseries}
\alias{find_timeseries}
\title{Identify Months With Overlapping Data}
\usage{
find_timeseries(
  isotopes,
  start_date = NULL,
  end_date = NULL,
  deep_ids = c("LL-01B", "LL-02B", "LL-05C", "LL-09B", "PFL-100", "PFL-101")
)
}
\arguments{
\item{isotopes}{a data frame with isotopes measurements as formatted in the
isotopes dataset, subset for a single lake.}

\item{start_date}{optional start date to use for analysis (first day of the
month) in POSIXct. If not provided, defaults to NULL and
timeseries is based on months with isotope samples.}

\item{end_date}{optional end date to use for analysis (last day of the
month) in POSIXct. If not provided, defaults to NULL and
timeseries is based on months with isotope samples.}

\item{deep_ids}{list of site_ids associated with deeps wells, which should
not be included when identifying start and end dates for
analysis purposes (if isotopes data frame is used to
determine start and end dates)}
}
\description{
Identifies the start date, end date, and number of months in the monthly
timeseries with complete input data coverage.
}
