% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_functions.R
\name{find_timeseries}
\alias{find_timeseries}
\title{Identify Months With Overlapping Data}
\usage{
find_timeseries(
  isotopes,
  lake_levels = NULL,
  gw_levels = NULL,
  dictionary = NULL,
  threshold = 0.01,
  by_gw_iso = FALSE
)
}
\arguments{
\item{isotopes}{a data frame with isotopes measurements as formatted in the
isotopes dataset, subset for a single lake.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to groundwater level records at the lake of
interest.}

\item{dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
dictionary dataset, subset to
records for the lake of interest.}

\item{threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in order
to classify groundwater measurement.}

\item{by_gw_iso}{logical defaults to TRUE to calculate water balance
30 days prior to upgradient groundwater isotope measurements.
If FALSE, calculates balance by calendar month instead (30
days before last day of the month)}
}
\description{
Identifies the start date, end date, and number of months in the monthly
timeseries with complete input data coverage.
}
