% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_d18O_evap.R
\name{get_monthly_d18O_evap}
\alias{get_monthly_d18O_evap}
\title{Monthly Evaporation d18O}
\usage{
get_monthly_d18O_evap(monthly_weather, monthly_lst, monthly_isotopes)
}
\arguments{
\item{monthly_weather}{data frame with the following columns:
\itemize{
\item date - first of the month for each monthly observation (Date)
\item atmp_K - mean monthly air temperature (K)
\item relh_pct - mean monthly relative humidity (percent)
\item pcpn_mm - monthly precipitaiton (mm)
\item rpet_mm - monthly reference potential evapotranspiration (mm)
}}

\item{monthly_lst}{a data frame with the following columns:
\itemize{
\item date - first of the month for each monthly observation (Date)
\item ltmp_K - mean monthly lake surface temperature (degrees K)
}}

\item{monthly_isotopes}{a data frame with the following columns:
\itemize{
\item date - first of the month for each monthly observation (Date)
\item d18O_pcpn - mean precipitation stable isotope measurement for the month
\item d18O_lake - mean lake stable isotope measurement for the month
\item d18O_GWin - mean groundwater inflow stable isotope measurement for the
                 month
}}
}
\value{
monthly_isotopes with d18O_evap column added
}
\description{
This function takes
}
