% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_values_functions.R
\name{get_monthly_dV}
\alias{get_monthly_dV}
\title{Monthly Change in Lake Volume}
\usage{
get_monthly_dV(lake_levels, stage_vol, month_info, as_depth = TRUE,
  tmp_depth_fix = FALSE)
}
\arguments{
\item{lake_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to lake level records for the lake of interest.}

\item{stage_vol}{a data frame with stage-volume-surface area relationships as
formatted in the \code{\link{stage_vol}} dataset, subset for
a single lake.}

\item{month_info}{a list with with the start date, end date, and number of
months in the common timeseries of input data, as output by
\code{\link{get_overlap_months}}}

\item{as_depth}{defaults to TRUE to normalize volume (m^3) by average surface
area of lake during the time segment (m^2) and return change
in volume as a depth (mm). If FALSE, returns change in lake
volume as a volume (m^3).}

\item{tmp_depth_fix}{defaults to FALSE. If TRUE, calculates change in volume
as a depth based solely on change in stage (mm).}
}
\value{
monthly_dV, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{dV}{change in lake volume during the month as a volume (m^3) or as a
          depth (mm), depending on value of as_depth}
}
}
\description{
Calculates the monthly change in lake volume based on daily lake levels and a
stage-volume relationship from bathymetry analysis. .
}
