% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_values_functions.R
\name{get_monthly_h2o_bal}
\alias{get_monthly_h2o_bal}
\title{Monthly Water Balance}
\usage{
get_monthly_h2o_bal(monthly_weather, monthly_isotopes, monthly_dV,
  month_info)
}
\arguments{
\item{monthly_weather}{a data frame as output by
\code{\link{get_monthly_weather}}}

\item{monthly_isotopes}{a data frame as output by
\code{\link{get_monthly_isotopes}}}

\item{monthly_dV}{a data frame as output by
\code{\link{get_monthly_dV}}}

\item{month_info}{a list with with the start date, end date, and number of
months in the common timeseries of input data, as output by
\code{\link{get_overlap_months}}}
}
\value{
monthly_h2o_bal, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{P}{monthly precipitation (mm)}
\item{E}{monthly evapotranspiration (mm)}
\item{GWin}{monthly groundwater inflow to the lake (mm)}
\item{GWout}{monthly groundwater outflow to the lake (mm)}
\item{dV}{monthly change in lake volume (mm)}
}
}
\description{
Calculates the water balance of a lake based measured fluxes and isotopic
signatures of precipitation, the lake, and inflowing groundwater wells. Based
on equations 2 and 4 in Krabbenhoft et al. (1990).
}
\references{
Krabbenhoft, D. P., C. J. Bowser, M. P. Anderson, and J. W.
  Valley. (1990). Estimating Groundwater Exchange with Lakes: 1. The Stable
  Isotope Mass Balance Method. Water Resources Research, 26(10):2445-2453.
  https://doi.org/10.1029/WR026i010p02445
}
