% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_isotopes.R
\name{get_monthly_isotopes}
\alias{get_monthly_isotopes}
\title{Get Monthly Isotopes}
\usage{
get_monthly_isotopes(lake, isotope_file,
  site_file = "csls_site_dictionary.csv", filedir = "system.file",
  Xday = 7)
}
\arguments{
\item{lake}{name of lake to analyze (capitalized, e.g. "Pleasant")}

\item{isotope_file}{name of isotope csv file with stable isotope measurement
                    info, including (but not limited to):
\itemize{
\item Site.ID - unique site id for where location was taken (e.g. LL-01 for
     Long Lake well 1). Assumed to match site ids in sites_file.
\item Valid - field indicating whether or not sample measurement is valid
      for analysis (TRUE) or not (FALSE).
\item d18O..VSMOW. - stable isotope measurement for d18O at this site
\item Collection.Date.Time - date and time of sample collection
}}

\item{site_file}{filname of site dictionary with the following columns
\itemize{
\item lake - name of lake associated with the measurement site. Corresponds
             to lake name argument in this function.
\item obs_type - type of observation (LK = lake, GW = groundwater
                 monitoring well)
\item site_id - unique site id for measurement site (e.g., LL-01).
                Corresponds to site_id in "isotopes" data frame.
\item SWIMS_station_id - SWIMS station id, if exists for this site.
\item USGS_id - USGS site number. Corresponds to site_no in "water_levels"
                data frame.
\item WBIC - water body identification code (WBIC), for lake sites only.
}}

\item{filedir}{name of directory with both csv files, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}

\item{Xday}{number of days to consider when calculating average water level
before isotope measurements, defaults to 7 days.}
}
\value{
monthly_isotopes, a data frame with the following:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{d18O_pcpn}{mean precipitation stable isotope measurement for the month}
\item{d18O_lake}{mean lake stable isotope measurement for the month}
\item{d18O_GWin}{mean groundwater inflow stable isotope measurement for the
                 month}
}
}
\description{
This function summarizes isotope measurements for precipitation, the lake,
and inflowing groundwater wells at a monthly time step. Requires one csv file
with stable isotope measurements and a second csv that acts as a key to the
relevant site ids for the lake of interest and also indicates what type of
sites they are (i.e., precipitaiton, lake, upstream groundwater well, or
downstream groundwater well). Groundwater inflow wells are determined a
priori based on visual inspection of groundwater levels in each well, lake
water level, and stable isotope measurements.
}
