% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_isotopes.R
\name{get_monthly_isotopes}
\alias{get_monthly_isotopes}
\title{Get Monthly Isotopes}
\usage{
get_monthly_isotopes(isotope_file, sites_file, filedir = "system.file")
}
\arguments{
\item{isotope_file}{name of isotope csv file with stable isotope measurement
                    info, including (but not limited to):
\itemize{
\item Site.ID - unique site id for where location was taken (e.g. LL-01 for
     Long Lake well 1). Assumed to match site ids in sites_file.
\item Valid - field indicating whether or not sample measurement is valid
      for analysis (TRUE) or not (FALSE).
\item d18O..VSMOW. - stable isotope measurement for d18O at this site
\item Collection.Date.Time - date and time of sample collection
}}

\item{sites_file}{name of csv file with info on the sampling sites for a
                  lake of interest, including:
\itemize{
\item site_id - unique site id for all stable isotope measurements relevant
      to this lake. Assumed to match site ids in isotope_file.
\item site_type - type of measurement location. Options include
      "precipitation", "lake", "upstream" (i.e., groundwater inflow well),
      "downstream" (i.e., groundwater outflow well), or "NA".
}}

\item{filedir}{name of directory with both csv files, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}
}
\value{
monthly_isotopes, a data frame with the following:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{d18O_pcpn}{mean precipitation stable isotope measurement for the month}
\item{d18O_lake}{mean lake stable isotope measurement for the month}
\item{d18O_GWin}{mean groundwater inflow stable isotope measurement for the
                 month}
}
}
\description{
This function summarizes isotope measurements for precipitation, the lake,
and inflowing groundwater wells at a monthly time step. Requires one csv file
with stable isotope measurements and a second csv that acts as a key to the
relevant site ids for the lake of interest and also indicates what type of
sites they are (i.e., precipitaiton, lake, upstream groundwater well, or
downstream groundwater well). Groundwater inflow wells are determined a
priori based on visual inspection of groundwater levels in each well, lake
water level, and stable isotope measurements.
}
