% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_values_functions.R
\name{get_monthly_isotopes}
\alias{get_monthly_isotopes}
\title{Monthly Isotope Measurements}
\usage{
get_monthly_isotopes(isotopes, site_dictionary, static_wells = TRUE,
  lake_levels = NULL, gw_levels = NULL, Xday = 7,
  static_lake = TRUE)
}
\arguments{
\item{isotopes}{a data frame with isotopes measurements as formatted in the
\code{\link{isotopes}} dataset, subset for a single lake.}

\item{site_dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
\code{\link{site_dictionary}} dataset, subset to
records for the lake of interest.}

\item{static_wells}{logical defaults to TRUE to use static definitions of
upgradient/downgradient wells in site dictionary. If
FALSE, uses lake_levels and gw_levels to define
upgradient/downgradient wells at each measurement date.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the \code{\link{water_levels}} dataset,
subset to groundwater level records at the lake of
interest.}

\item{Xday}{number of days to consider when calculating average water level
before isotope measurements, defaults to 7 days.}

\item{static_lake}{logical defaults to TRUE to use mean of fall (Sept-Nov)
isotope samples for the lake. If FALSE, uses actual
measurement for each month.}
}
\value{
monthly_isotopes, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{d18O_pcpn}{mean precipitation stable isotope measurement for the month}
\item{d18O_lake}{mean lake stable isotope measurement for the month}
\item{d18O_GWin}{mean groundwater inflow stable isotope measurement for the
                 month}
\item{d18O_GWout}{mean groundwater outflow stable isotope measurement for the
                 month}
}
}
\description{
Summarizes isotope measurements at a monthly timestep for a single lake
}
\seealso{
\code{\link{retrieve_csls_isotopes}}, \code{\link{isotopes}},
         \code{\link{retrieve_csls_water_levels}},
         \code{\link{water_levels}},
         \code{\link{retrieve_csls_site_dictionary}},
         \code{\link{site_dictionary}}
}
