% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_lst.R
\name{get_monthly_lst}
\alias{get_monthly_lst}
\title{Monthly Lake Surface Temperature}
\usage{
get_monthly_lst(filename, filedir = "system.file", wbic = 106900)
}
\arguments{
\item{filename}{name of SWIMS csv file which includes the following fields:
\itemize{
  \item DNR Parameter - unique DNR number for parameter. Assume lake
        surface temperatures are stored as "10" for "Temperature Field"
  \item Result - numeric value of measurement
  \item Units - units of result
  \item Result Depth - depth at which measurement taken, asume is "0
        Meters" for lake surface temperatures
  \item WBIC - Wisconsin Water Body Identification Code
}}

\item{filedir}{directory in which daily weather csv file resides, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}

\item{wbic}{Wisconsin Water Body Identification Code for lake of interest}
}
\value{
monthly_lst, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{ltmp}{mean monthly lake surface temperature (degrees C)}
}
}
\description{
This function takes output from the Wisconsin Department of Natural
Resources Surface Water Integrated Monitoring System (SWIMS),
}
\references{
\url{https://dnr.wi.gov/topic/surfacewater/swims/}
}
