% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_values_functions.R
\name{get_monthly_lst}
\alias{get_monthly_lst}
\title{Monthly Lake Surface Temperature}
\usage{
get_monthly_lst(lst)
}
\arguments{
\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the \code{\link{lst}} dataset, subset
for a single lake.}
}
\value{
monthly_lst, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{ltmp_K}{mean monthly lake surface temperature (degrees K)}
}
}
\description{
Summarizes sub-monthly lake surface temperature at a monthly timestep for a
single lake
}
\seealso{
\code{\link{retrieve_csls_lst}}, \code{\link{lst}}
}
