% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_weather.R
\name{get_monthly_weather}
\alias{get_monthly_weather}
\title{Get Monthly Weather}
\usage{
get_monthly_weather(filename, filedir = "system.file", skip_lines = 7,
  daily_weather = FALSE)
}
\arguments{
\item{filename}{name of weather csv file which includes:
\itemize{
  \item date (m/d/YYYY)
  \item (atmp_max and atmp_min) OR atmp - min/max air temperature (daily, deg
        C) OR air temperature (subdaily, deg C)
  \item (relh_max and relh_min) OR relh - min/max relative humidity (daily,
        %) OR relative humidity (subdaily, %)
  \item pcpn - precipitation (mm)
  \item rpet - reference potential evapotranspiration (mm)
}}

\item{filedir}{directory in which weather csv file resides, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory.}

\item{skip_lines}{[numeric] number of rows in csv to skip before reading in
data, defaults to 7.}

\item{daily_weather}{[logical] defaults to FALSE for subdaily weather, set
to TRUE for daily weather}
}
\value{
monthly_weather, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{atmp_K}{mean monthly air temperature (K)}
\item{relh_pct}{mean monthly relative humidity (percent)}
\item{pcpn_mm}{monthly precipitaiton (mm)}
\item{rpet_mm}{monthly reference potential evapotranspiration (mm)}
}
}
\description{
This function summarizes daily or subdaily weather at a monthly timestep.
Required weather fields include air temperature (min/max if daily, deg C),
relative humidity (min/max if daily, %), precipitation (mm), and reference
evapotranspiration (mm). Assumes column names and units are as received from
the Hancock, WI weather station.
}
\references{
\url{https://enviroweather.msu.edu/weather.php?stn=hck}
}
