% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_values_functions.R
\name{get_monthly_weather}
\alias{get_monthly_weather}
\title{Monthly Weather}
\usage{
get_monthly_weather(weather)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather as formatted in the
\code{\link{weather}} dataset.}
}
\value{
monthly_weather, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{atmp_K}{mean monthly air temperature (K)}
\item{relh_pct}{mean monthly relative humidity (percent)}
\item{pcpn_mm}{monthly precipitaiton (mm)}
\item{rpet_mm}{monthly reference potential evapotranspiration (mm)}
}
}
\description{
Summarizes sub-monthly weather at a monthly timestep
}
\seealso{
\code{\link{retrieve_csls_weather}}, \code{\link{weather}}
}
