% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_site_type.R
\name{iso_site_type}
\alias{iso_site_type}
\title{Classify Groundwater Stable Isotope Measurements (Dynamic)}
\usage{
iso_site_type(isotopes, dictionary, static_gw = FALSE,
  lake_levels = NULL, gw_levels = NULL, threshold = 0.01)
}
\arguments{
\item{isotopes}{a data frame with isotopes measurements as formatted in the
isotopes dataset, subset for a single lake.}

\item{dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
dictionary dataset, subset to
records for the lake of interest.}

\item{static_gw}{logical defaults to FALSE to use lake_levels and gw_levels
to define upgradient/downgradient wells at each measurement
date. If TRUE, uses static definitions of
upgradient/downgradient wells in site dictionary.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to groundwater level records at the lake of
interest.}

\item{threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in order
to classify groundwater measurement.}
}
\description{
Classifies the site type of isotope measurements as precipitation, lake,
upgradient, or downgradient. Upgradient vs. downgradient wells are determined
by comparing the daily difference in lake levels and well levels in the 30
days leading up to isotope measurements.
}
