% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_functions.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot Balance}
\usage{
plot_balance(
  lake,
  monthly_h2o_bal,
  as_vol = TRUE,
  annual = FALSE,
  text_size = 12
)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{monthly_h2o_bal}{a data frame with the date and all fluxes into and out
of the lake.}

\item{as_vol}{logical defaults to TRUE to plot water balance as a volume
(m^3). If FALSE, plots as depths.}

\item{annual}{defaults to FALSE to indicate monthly timestep. Set to TRUE to
indicate annual balance.}

\item{text_size}{size of font, defaults to 12 point}
}
\value{
plot_obj - a plot object with aesthetics added
}
\description{
This function creates a plot object to evaluate the timeseries of d18O
measurements over time at each site.
}
