% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runall_csls_budget.R
\name{runall_csls_budget}
\alias{runall_csls_budget}
\title{Run isoH2Obudget start to finish for a CSLS lake}
\usage{
runall_csls_budget(lake, threshold = 0.01, by_gw_iso = FALSE,
  annual = FALSE)
}
\arguments{
\item{lake}{name of lake (e.g., Pleasant, Long, or Plainfield)}

\item{threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in order
to classify groundwater measurement.}

\item{by_gw_iso}{logical defaults to TRUE to calculate water balance
30 days prior to upgradient groundwater isotope measurements.
If FALSE, calculates balance by calendar month instead (30
days before last day of the month)}

\item{annual}{defaults to FALSE to calculate water balance on a monthly
basis. If TRUE, calculates the annual balance instead.}
}
\value{
monthly_h2o_bal, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{P}{monthly precipitation (mm)}
\item{E}{monthly evapotranspiration (mm)}
\item{GWin}{monthly groundwater inflow to the lake (mm)}
\item{GWout}{monthly groundwater outflow to the lake (mm)}
\item{dV}{monthly change in lake volume (mm)}
}
}
\description{
Runs all major functions in the isoH2Obudget package, from loading in raw
data to the final monthly water balance for one of the CSLS lakes.
}
\details{
Loaded data includes the following data in a list format, where each lake has
its own data frame. For example, to access groundwater levels for Pleasant
Lake, the command would be \code{CSLSdata::gw_levels[["Pleasant"]]}.
\itemize{
\item \code{\link[CSLSdata]{weather}}, hourly weather data including "date",
"atmp" (air temperature, deg C), "P" (precipitation, mm), "RH"
(relative humidity, percent), "Rs" (solar radiation, MJ), and "wind"
(wind speed, m/s)
\item \code{\link[CSLSdata]{lst_HOBO}}, hourly lake surface temperature from
HOBO loggers, including "date", "ltmp" (lake surface temperature, deg
C), and "elev_m" (elevation of the sensor, meters above mean sea level)
\item \code{\link[CSLSdata]{isotopes}}, stable isotope measurements including
"date", "lake", "site_id", "d18O" measurement, and "d2H" measurement
\item \code{\link[CSLSdata]{lake_levels}}, daily lake levels (meters above
mean sea level) including "date", "
\item \code{\link[CSLSdata]{gw_levels}}, daily groundwater levels including
date, site_no, obs_type ("GW"), level_m (level in meters above mean sea
level), and site_id (CSLS site id)
\item \code{\link[CSLSdata]{dictionary}}, links site ids and characteristics
together with columns for lake, obs_type ("LK", "GW", "P"), site_id
(e.g., "LL-01"), SWIMS_station_id, USGS_id, WBIC, static_iso_class
(e.g., "upgradient", "lake", "precipitation"), lat_deg, long_deg,
elev_m, and bouy_bottom_elev_m
}
}
