% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runall_isoH2Obudget.R
\name{runall_isoH2Obudget}
\alias{runall_isoH2Obudget}
\title{Run isoH2Obudget Start to Finish}
\usage{
runall_isoH2Obudget(lake, files, filedir = "system.file",
  dataset = "csls")
}
\arguments{
\item{lake}{name of lake (e.g., Pleasant, Long, or Plainfield)}

\item{files}{a list that includes the following filenames:
\itemize{
\item weather_file
\item SWIMS_file
\item isotope_file
\item site_file
\item stage_vol_file
}}

\item{filedir}{name of directory with both csv files, defaults to
'system.file' to instruct it to look within package
"inst/extdata" directory}

\item{dataset}{dataset type to process. Expects "csls" or "kniffin".}
}
\value{
monthly_h2o_bal, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation (Date)}
\item{P}{monthly precipitation (mm)}
\item{E}{monthly evapotranspiration (mm)}
\item{GWin}{monthly groundwater inflow to the lake (mm)}
\item{GWout}{monthly groundwater outflow to the lake (mm)}
\item{dV}{monthly change in lake volume (mm)}
}
}
\description{
Runs all major functions in the isoH2Obudget package, from loading in raw
data to the final monthly water balance
}
\examples{
\dontrun{
csls_files <- list(weather_file = "csls_weather_hancock_hourly_2019_09_30.csv",
                   SWIMS_file = "csls_SWIMS_2019_09_05.csv",
                   isotope_file = "csls_isotope_samples_2019_10_02.csv",
                   site_file = "csls_site_dictionary.csv",
                   stage_vol_file = "csls_stage_volume.csv")
lake = "Pleasant"
monthly_h2o_bal <- runall_isoH2Obudget(lake, csls_files,
                                       filedir = "system.file",
                                       dataset = "csls")


kniffin_files <- list(weather_file = "kniffin_weather_monthly.csv",
                      isotope_file = "kniffin_isotope_samples.csv",
                      site_file = "kniffin_site_dictionary.csv",
                      d18O_evap_file = "kniffin_d18O_evap.csv")
lake = "Long"
monthly_h2o_bal <- runall_isoH2Obudget(lake, kniffin_files,
                                       filedir = "system.file",
                                       dataset = "kniffin")
}
}
