% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_lake_functions.R
\name{subset_lake}
\alias{subset_lake}
\title{Subset Inputs for Given Lake}
\usage{
subset_lake(lake, site_dictionary, lst, isotopes, water_levels, stage_vol)
}
\arguments{
\item{lake}{lake of interest (e.g., "Pleasant", "Long", or "Plainfield")}

\item{site_dictionary}{a data frame with the lake, obs_type, site_id,
SWIMS_station, USGS_id, WBIC, and static_iso_type of
measurement sites, as in the
\code{\link{site_dictionary}} dataset}

\item{lst}{a data frame with the WBIC, date, ltmp, and units of
every-other-week lake surface temperature measurements, as in the
\code{\link{lst}} dataset}

\item{isotopes}{a data frame with the date, lake, site_id, d18O, and d2H of
isotope measurements, as in the \code{\link{isotopes}}
dataset}

\item{water_levels}{a data frame with the date, site_no, obs_type, and
level_m of daily water level observations, as in the
\code{\link{water_levels}} dataset.}

\item{stage_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the \code{\link{stage_vol}} dataset.}
}
\value{
a list with the input data frames subsetted to measurements
  associated with the lake of interest. Includes:
\describe{
\item{lst}{data frame with lake surface temperatures (deg C)}
\item{isotopes}{data frame with lake isotope measurements, including
                precipitation measurements}
\item{lake_levels}{data frame with lake levels (m)}
\item{gw_levels}{data frame with groundwater levels (m) around the lake}
\item{stage_vol}{data frame with relationships among stage (m), volume (m^3),
                 and surface area (m^2) for the lake}
\item{site_dictionary}{data frame with all names, identification codes, and
                       classifications of measurment sites associated with
                       the lake}
}
}
\description{
This function subsets input data for isoH2Obudget for just the lake of
interest. Includes: site_dictionary, lst, isotopes, water_levels (to
gw_levels and lake_levels), and stage_vol.
}
