% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dV.R
\name{summarise_dV}
\alias{summarise_dV}
\title{Monthly Change in Lake Volume}
\usage{
summarise_dV(lake_levels, timeseries)
}
\arguments{
\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{timeseries}{a list with dates (POSIXct) to analyze, and intervals
(lubridate interval) covering the month before each analysis
date.}
}
\value{
monthly_dV, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{dV}{change in lake volume during the month (m^3)}
\item{mean_vol_m3}{mean lake volume during the month (m^3)}
\item{mean_area_m2}{mean lake surface area during the month (m^2)}
}
}
\description{
Calculates the monthly change in lake volume based on daily lake levels and
an elevation-volume relationship from bathymetry analysis.
}
