% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_inputs.R
\name{summarise_inputs}
\alias{summarise_inputs}
\title{Summarize Inputs over Common Timeseries}
\usage{
summarise_inputs(weather, lst, isotopes, lake_levels, gw_levels,
  elev_area_vol, dictionary, threshold = 0.01, by_gw_iso = FALSE,
  annual = FALSE)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including including date
(POSIXct), atmp (air temperature, deg C), P (precipitation, mm), RH
(relative humidity, percent), Rs (solar radiation, MJ), and wind (wind
speed, m/s)}

\item{lst}{a data frame with sub-monthly lake surface temperature
measurements as formatted in the lst_HOBO dataset, subset
for a single lake.}

\item{isotopes}{a data frame with isotopes measurements as formatted in the
isotopes dataset, subset for a single lake.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to groundwater level records at the lake of
interest.}

\item{elev_area_vol}{a data frame with the lake, stage_m, surf_area_m2, and
volume_m3 as in the elev_area_voldataset, subset
for a single lake.}

\item{dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
dictionary dataset, subset to
records for the lake of interest.}

\item{threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in order
to classify groundwater measurement.}

\item{by_gw_iso}{logical defaults to TRUE to calculate water balance
30 days prior to upgradient groundwater isotope measurements.
If FALSE, calculates balance by calendar month instead (30
days before last day of the month)}

\item{annual}{defaults to FALSE to calculate water balance on a monthly
basis. If TRUE, calculates the annual balance instead.}
}
\value{
h2o_bal_inputs, a data frame ordered by date with data only for the
        overlap timeseries for:
\describe{date}{dates of monthly timeseries (1st of the month) for months with
                complete overlapping input data}
\describe{atmp_degC}{mean air temperature (degrees C)}
\describe{RH_pct}{mean relative humidity (percent)}
\describe{P_mm}{total precipitation (mm)}
\describe{ET_mm}{total lake evaporation (mm)}
\describe{atmp_K}{mean air temperature (Kelvin)}
\describe{ltmp_K}{mean lake surface temperature (degrees C)}
\describe{dV}{change in lake volume (mm)}
\describe{d18O_lake}{mean stable isotope measurement for lake}
\describe{d18O_pcpn}{mean stable isotope measurement for precipitation}
\describe{d18O_GWin}{mean stable isotope measurement for inflowing
                     groundwater}
\describe{d18O_GWout}{mean stable isotope measurement for outflowing
                      groundwater}
\describe{d18O_evap}{mean stable isotope measurement for evaporation}
\describe{GWin_sites}{wells included in mean inflowing groundwater stable
                      isotope measurement}
\describe{GWout_sites}{wells included in mean outflowing groundwater stable
                      isotope measurement}
}
\description{
This function takes raw weather, lake surface temperature, isotope, lake
level, groundwater level, stage-volume inputs and summarises water balance
inputs over a common timeseries.
}
