% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_isotopes.R
\name{summarise_isotopes}
\alias{summarise_isotopes}
\title{Monthly Isotope Measurements}
\usage{
summarise_isotopes(isotopes, site_dictionary, timeseries,
  static_gw = FALSE, lake_levels = NULL, gw_levels = NULL,
  median_threshold = 0.01, static_lake = FALSE,
  use_kniffin_pcpn = TRUE,
  pcpnfile = "csls_isotope_precipitation_deployment.csv",
  pcpndir = "system.file")
}
\arguments{
\item{isotopes}{a data frame with isotopes measurements as formatted in the
isotopes dataset, subset for a single lake.}

\item{site_dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
site_dictionary dataset, subset to
records for the lake of interest.}

\item{timeseries}{a vector of all months in the common timeseries among input
datasets}

\item{static_gw}{logical defaults to FALSE to use lake_levels and gw_levels
to define upgradient/downgradient wells at each measurement
date. If TRUE, uses static definitions of
upgradient/downgradient wells in site dictionary.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the water_levels dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the water_levels dataset,
subset to groundwater level records at the lake of
interest.}

\item{median_threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in
order to classify groundwater measurement.}

\item{static_lake}{logical defaults to FALSE to use actual measurement for
each month. If TRUE, uses mean of fall (Sept-Nov) isotope
samples for the lake.}

\item{use_kniffin_pcpn}{logical defaults to TRUE to average in precipitation
measurements by Maribeth Kniffin for each month.}

\item{pcpnfile}{name of file with dates of precipitation collector deployment.
Defaults to "csls_isotope_precipitation_deployment.csv"}

\item{pcpndir}{name of directory with file with dates of precipitation
collector deployment. Defaults to "system.file" to indicate is
stored within inst/extdata within the installed isoH2Obudget
package files.}
}
\value{
monthly_isotopes, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{d18O_pcpn}{mean precipitation stable isotope measurement for the month}
\item{d18O_lake}{mean lake stable isotope measurement for the month}
\item{d18O_GWin}{mean groundwater inflow stable isotope measurement for the
                 month}
\item{d18O_GWout}{mean groundwater outflow stable isotope measurement for the
                 month}
\item{GWin_sites}{well sites used in mean d18O_GWin each month}
\item{GWout_sites}{well sites used in mean d18O_GWout each month}
}
}
\description{
Summarizes isotope measurements at a monthly timestep for a single lake
}
