% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_isotopes.R
\name{summarise_isotopes}
\alias{summarise_isotopes}
\title{Monthly Isotope Measurements}
\usage{
summarise_isotopes(isotopes, dictionary, timeseries, lake_levels,
  gw_levels, threshold = 0.01)
}
\arguments{
\item{isotopes}{a data frame with isotopes measurements as formatted in the
isotopes dataset, subset for a single lake.}

\item{dictionary}{a data frame with site id numbers and static isotope
site classifications as formatted in the
dictionary dataset, subset to
records for the lake of interest.}

\item{timeseries}{a list with dates (POSIXct) to analyze, and intervals
(lubridate interval) covering the month before each analysis
date.}

\item{lake_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to lake level records for the lake of interest.}

\item{gw_levels}{a data frame with daily water level measurements as
formatted in the lake_levels dataset,
subset to groundwater level records at the lake of
interest.}

\item{threshold}{minimum median difference between lake levels and
groundwater levels during the month of measurement in order
to classify groundwater measurement.}
}
\value{
monthly_isotopes, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{d18O_pcpn}{mean precipitation stable isotope measurement for the month}
\item{d18O_lake}{mean lake stable isotope measurement for the month}
\item{d18O_GWin}{mean groundwater inflow stable isotope measurement for the
                 month}
\item{d18O_GWout}{mean groundwater outflow stable isotope measurement for the
                 month}
\item{GWin_sites}{well sites used in mean d18O_GWin each month}
\item{GWout_sites}{well sites used in mean d18O_GWout each month}
}
}
\description{
Summarizes isotope measurements at a monthly timestep for a single lake
}
