% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_weather.R
\name{summarise_weather}
\alias{summarise_weather}
\title{Monthly Weather}
\usage{
summarise_weather(weather, timeseries, lake)
}
\arguments{
\item{weather}{a data frame with sub-monthly weather including including date
(POSIXct), atmp (air temperature, deg C), P (precipitation, mm), RH
(relative humidity, percent), Rs (solar radiation, MJ), and wind (wind
speed, m/s)}

\item{timeseries}{a list with dates (POSIXct) to analyze, and intervals
(lubridate interval) covering the month before each analysis
date.}

\item{lake}{name of lake (e.g., Pleasant, Long, or Plainfield)}
}
\value{
monthly_weather, a data frame with the following columns:
\describe{
\item{date}{first of the month for each monthly observation}
\item{atmp_K}{mean monthly air temperature (K)}
\item{RH_pct}{mean monthly relative humidity (percent)}
\item{P_mm}{monthly precipitaiton (mm)}
\item{ET_mm}{monthly lake evaporation (mm)}
}
}
\description{
Summarizes sub-monthly weather at a monthly timestep.
}
